% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add missing metadata to a TIRTLseqData object.}
\usage{
add_metadata(obj, sep = "_", meta_columns)
}
\description{
Add missing metadata to a TIRTLseqData object.
}
\seealso{
Other data_wrangling: 
\code{\link{filter_dataset}()},
\code{\link{load_tirtlseq}()},
\code{\link{reorder_samples}()}
}
\concept{data_wrangling}
