% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_add_single_chain_data.R
\name{add_single_chain_data}
\alias{add_single_chain_data}
\title{Add single-chain read counts/fractions to the paired TCR data}
\usage{
add_single_chain_data(data, verbose = TRUE)
}
\arguments{
\item{data}{a TIRTLseq dataset created by \code{\link{load_tirtlseq}()}}

\item{verbose}{(optional) whether to print progress of the function (default is TRUE).}
}
\value{
A TIRTLseq dataset object where the paired data frames for each sample have
added columns for read counts and proportions from the single-chain pseudobulk data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function has been deprecated because its functionality has been added to the load_tirtlseq() function

The \code{add_single_chain_data()} function adds read counts and proportions
from the single-chain pseudobulk data to the paired data frame for each sample
of a dataset.
}
\seealso{
Other deprecated:
\code{\link[=identify_non_functional_seqs]{identify_non_functional_seqs()}},
\code{\link[=identify_paired]{identify_paired()}},
\code{\link[=process_TIRTLseq]{process_TIRTLseq()}}
}
\concept{deprecated}
