% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pseudobulk_counts_to_paired_data.R
\name{add_single_chain_data_simple}
\alias{add_single_chain_data_simple}
\title{The input 'data' must be a list with 3 items, "alpha", "beta", and "paired"
This looks up readCounts/Fractions of each single chain in the}
\usage{
add_single_chain_data_simple(data)
}
\description{
The input 'data' must be a list with 3 items, "alpha", "beta", and "paired"
This looks up readCounts/Fractions of each single chain in the
}
