% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_pair_manual.R
\name{choose_pair_manual}
\alias{choose_pair_manual}
\title{Choose a partner manually for an input chain using T-SHELL}
\usage{
choose_pair_manual(res, rank)
}
\arguments{
\item{res}{an object created by \code{\link{plot_tshell}()}}

\item{rank}{the rank of the potential partner chain to choose}
}
\value{
A one-row data frame in the style of the "paired" data frame from pairing output
with the input chain and its chosen partner.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function takes in an object created by \code{\link{plot_tshell}()} and allows
a user to choose the partner for the input chain used to create that object.
}
\seealso{
Other well: 
\code{\link{get_well_subset}()},
\code{\link{get_wells_from_edges}()},
\code{\link{load_well_counts_binary}()},
\code{\link{plot_tshell}()},
\code{\link{write_well_data_to_binary}()}
}
\concept{well}
