% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_scTIRTLseq.R
\name{clean_scTIRTLseq}
\alias{clean_scTIRTLseq}
\title{Imputes missing alpha and beta chains where possible for single-cell TIRTLseq data}
\usage{
clean_scTIRTLseq(df, verbose = TRUE, keep_all_columns = FALSE)
}
\description{
Imputes missing alpha and beta chains where possible for single-cell TIRTLseq data
}
\seealso{
Other single-cell: 
\code{\link{summarize_scTIRTLseq}()}
}
\concept{single-cell}
