% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_thimble_df.R
\name{create_thimble_df}
\alias{create_thimble_df}
\title{Convert paired TCRs to a "thimble" data frame for stitching together TCR nucleotide coding sequences with the Stitchr Python package}
\usage{
create_thimble_df(
  df,
  preset = c("default", "none"),
  TCR_names = "TCR",
  exclude_non_functional = TRUE,
  remove_duplicates = TRUE,
  verbose = TRUE,
  Linker = NULL,
  Link_order = NULL,
  TRAC = NULL,
  TRBC = NULL,
  TRA_leader = NULL,
  TRB_leader = NULL,
  TRA_5_prime_seq = NULL,
  TRA_3_prime_seq = NULL,
  TRB_5_prime_seq = NULL,
  TRB_3_prime_seq = NULL
)
}
\description{
Convert paired TCRs to a "thimble" data frame for stitching together TCR nucleotide coding sequences with the Stitchr Python package
}
\seealso{
Other stitchr: 
\code{\link{write_stitchr_tsv}()}
}
\concept{stitchr}
