% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download TIRTLseq datasets}
\usage{
download_dataset(dataset = "SJTRC_TIRTLseq_minimal")
}
\description{
Download TIRTLseq datasets
}
\concept{download}
