% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{filter_dataset}
\alias{filter_dataset}
\title{Subset a TIRTLseqData object}
\usage{
filter_dataset(data, samples)
}
\arguments{
\item{data}{a TIRTLseqData object created by \code{\link{load_tirtlseq}()}}

\item{samples}{the selected samples. Either 1) a numeric vector of indices,
2) a character vector of sample names, or 3) a character vector of metadata
conditions where each entry is of the form \code{"column==value"}.}
}
\value{
A dataset object similar to that created by \code{\link{load_tirtlseq}()},
but with only the selected samples.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The \code{filter_dataset()} function is used to select a subset of samples from
a loaded TIRTLseq dataset and create a new dataset object.
}
\details{
The function accepts 1) a numeric vector of indices, 2) a character vector of sample names, or
3) a character vector of metadata conditions where each entry is of the form \code{"column==value"}.

In the third case, \code{c("cell_type==cd4", "timepoint==tp2")} would, for example, select
samples whose \code{cell_type} is \code{cd4} and whose \code{timepoint} is \code{tp2} in
the sample metadata.
}
\examples{
folder = system.file("extdata/SJTRC_TIRTL_seq_longitudinal",
  package = "TIRTLtools")
sjtrc = load_tirtlseq(folder,
  meta_columns = c("marker", "timepoint", "version"), sep = "_",
  chain = "paired", verbose = FALSE)
p2 = filter_dataset(sjtrc, 1:3) ### by indices
p3 = filter_dataset(sjtrc, c("cd8_tp1_v2", "cd8_tp2_v2", "cd8_tp3_v2")) ### by sample names
p4 = filter_dataset(sjtrc, "marker==cd4") ### by sample metadata condition
p5 = filter_dataset(sjtrc, c("marker==cd4", "timepoint==tp2")) ### by multiple sample metadata conditions

}
\seealso{
Other data_wrangling: 
\code{\link{add_metadata}()},
\code{\link{reorder_samples}()}
}
\concept{data_wrangling}
