% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicate_tcrs.R
\name{filter_duplicate_tcrs}
\alias{filter_duplicate_tcrs}
\title{Remove duplicate TCRs from a data frame}
\usage{
filter_duplicate_tcrs(df, verbose = TRUE)
}
\arguments{
\item{df}{data frame with paired TCRs}

\item{verbose}{whether to print number of TCRs removed}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

By default, the TIRTLseq pairing scripts write a "paired" data table that has one row
per pair per pairing algorithm (MAD-HYPE or T-SHELL). Thus, any pairs identified by both methods
will be found in two rows. This function removes those duplicates.
}
\seealso{
Other data_processing:
\code{\link[=clean_pairs]{clean_pairs()}},
\code{\link[=combine_bulk_and_paired_data]{combine_bulk_and_paired_data()}},
\code{\link[=filter_mait]{filter_mait()}},
\code{\link[=filter_nonfunctional_TCRs]{filter_nonfunctional_TCRs()}},
\code{\link[=filter_short_cdr3s]{filter_short_cdr3s()}},
\code{\link[=filter_v_alleles]{filter_v_alleles()}},
\code{\link[=make_tcr_schema]{make_tcr_schema()}},
\code{\link[=prep_for_tcrdist]{prep_for_tcrdist()}},
\code{\link[=remove_duplicates]{remove_duplicates()}}
}
\concept{data_processing}
