% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nonfunctional_TCRs.R
\name{filter_nonfunctional_TCRs}
\alias{filter_nonfunctional_TCRs}
\title{Remove TCRs with nonfunctional CDR3 amino acid sequences}
\usage{
filter_nonfunctional_TCRs(data, verbose = TRUE)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{verbose}{whether to print number of TCRs removed}
}
\value{
Returns a TIRTLseqDataSet object with nonfunctional TCRs removed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function removes TCR pairs where the CDR3-alpha or beta segment does not make a
functional protein, i.e. amino acid sequence contains a stop codon (*) or a frameshift (_).
}
\seealso{
Other data_processing:
\code{\link[=clean_pairs]{clean_pairs()}},
\code{\link[=combine_bulk_and_paired_data]{combine_bulk_and_paired_data()}},
\code{\link[=filter_duplicate_tcrs]{filter_duplicate_tcrs()}},
\code{\link[=filter_mait]{filter_mait()}},
\code{\link[=filter_short_cdr3s]{filter_short_cdr3s()}},
\code{\link[=filter_v_alleles]{filter_v_alleles()}},
\code{\link[=make_tcr_schema]{make_tcr_schema()}},
\code{\link[=prep_for_tcrdist]{prep_for_tcrdist()}},
\code{\link[=remove_duplicates]{remove_duplicates()}}
}
\concept{data_processing}
