% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_well_data.R
\name{filter_well_data}
\alias{filter_well_data}
\title{Remove rare clones from well-level TIRTLseq data}
\usage{
filter_well_data(well_data_df, min_fraction = 1e-06)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other well: 
\code{\link{get_well_subset}()},
\code{\link{get_wells_from_edges}()},
\code{\link{load_well_data}()},
\code{\link{load_well_data_to_sparse_multi}()}
}
\concept{well}
