% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{get_all_div_metrics}
\alias{get_all_div_metrics}
\title{Get all available diversity metrics}
\usage{
get_all_div_metrics()
}
\value{
A vector of all available diversity metrics for the \code{\link{diversity}()} function.
}
\description{
Get all available diversity metrics
}
\examples{
# example code
get_all_div_metrics()

}
\seealso{
\code{\link{diversity}()}, \code{\link{plot_diversity}()}
}
