% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paired_by_read_fraction_range.R
\name{get_paired_by_read_fraction_range}
\alias{get_paired_by_read_fraction_range}
\title{Calculate the number and fraction of single chains that were paired by frequency}
\usage{
get_paired_by_read_fraction_range(
  data,
  chain = c("beta", "alpha"),
  cutoffs = 10^(-6:-1),
  by_method = TRUE
)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{chain}{the TCR chain to use for read fraction (default is "beta")}

\item{cutoffs}{a vector of cutoffs for the read fraction ranges}

\item{by_method}{(optional) whether to get stats for each pairing method}
}
\value{
a data frame with the number and fraction of chains paired for each
read fraction range for each sample.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function returns the number and fraction of single-chains (alpha or beta, default is beta)
that were paired within different frequency ranges
(default is \verb{[10^-1, 10^-2], [10^-2, 10^-3], ... , [10^-5,10^-6], [10^-6, 0]}) for each sample.
You can use this function to get the data used in the plot for
\code{\link{plot_paired_by_read_fraction_range}()}.
}
\examples{
load_example_data(dataset = "SJTRC_longitudinal")

get_paired_by_read_fraction_range(SJTRC_longitudinal, chain = "beta")


}
\seealso{
Other qc:
\code{\link[=get_all_tcrs]{get_all_tcrs()}},
\code{\link[=get_pair_stats]{get_pair_stats()}},
\code{\link[=plot_n_paired_step]{plot_n_paired_step()}},
\code{\link[=plot_n_reads]{plot_n_reads()}},
\code{\link[=plot_num_partners]{plot_num_partners()}},
\code{\link[=plot_paired]{plot_paired()}},
\code{\link[=plot_paired_by_read_fraction_range]{plot_paired_by_read_fraction_range()}},
\code{\link[=plot_paired_vs_rank]{plot_paired_vs_rank()}},
\code{\link[=plot_pairs_with_eachother]{plot_pairs_with_eachother()}},
\code{\link[=plot_ranks]{plot_ranks()}},
\code{\link[=plot_read_fraction_vs_pair_status]{plot_read_fraction_vs_pair_status()}},
\code{\link[=plot_sample_overlap]{plot_sample_overlap()}},
\code{\link[=qc_report]{qc_report()}},
\code{\link[=summarize_data]{summarize_data()}}
}
\concept{qc}
