% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{get_samples}
\alias{get_samples}
\title{Returns sample names that are available to be imported from a directory}
\usage{
get_samples(data_dirs, pattern = NULL)
}
\arguments{
\item{data_dirs}{a directory path (or a vector with multiple paths) with TIRTL-seq data files}

\item{pattern}{a filename pattern}
}
\value{
A dataframe with a column for the sample name and other columns to indicate
if the sample has paired TCR data, TCRalpha pseudobulk data, and/or TCRbeta pseudobulk data.
}
\description{
The \code{get_samples()} takes a directory path and searches the directory for
TIRTL-seq files -- files ending in \verb{_TIRTLoutput.tsv(.gz)}, \verb{_pseudobulk_TRA.tsv(.gz)},
or \verb{_pseudobulk_TRB.tsv(.gz)}. You can give a regex pattern to return only
samples that match this pattern.
}
\seealso{
Other data_import:
\code{\link[=load_example_data]{load_example_data()}},
\code{\link[=load_tirtlseq]{load_tirtlseq()}},
\code{\link[=read_external_bulk]{read_external_bulk()}},
\code{\link[=read_external_paired]{read_external_paired()}}
}
\concept{data_import}
