% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{get_samples_single_dir}
\alias{get_samples_single_dir}
\title{Internal helper for \code{get_samples()}}
\usage{
get_samples_single_dir(data_dir, pattern = NULL)
}
\description{
same as \code{get_samples()}, but it takes in one directory only
}
\keyword{internal}
