% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wells_from_edges.R
\name{get_wells_from_edges}
\alias{get_wells_from_edges}
\title{Get well names from the top left and bottom right wells}
\usage{
get_wells_from_edges(
  top_left,
  bottom_right,
  return_type = c("wells", "rows_and_columns")
)
}
\arguments{
\item{top_left}{the well in the top left of the rectangle (e.g. "A1")}

\item{bottom_right}{the well in the top left of the rectangle (e.g. "H12")}
}
\value{
If return_type = "wells" (default), returns a vector of well names
covering the specified rows and columns (e.g. "A1", "A2", ... "H12"). If
return_type = "rows_and_columns", returns a list with two vectors
\itemize{
\item rows - a vector of the rows (e.g. "A", "B", ..., "H")
\item columns - a vector of the columns (e.g. 1, 2, 3, ..., 24)
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function returns a vector of well names in a rectangle containing the top
left well and the bottom right well.
}
\examples{
get_wells_from_edges("A13", "P24") ## right half of plate

get_wells_from_edges("A1", "H12") ## top left half of plate

get_wells_from_edges("A13", "P24", return_type = "rows_and_columns")

}
\seealso{
Other well: 
\code{\link{choose_pair_manual}()},
\code{\link{get_well_subset}()},
\code{\link{load_well_counts_binary}()},
\code{\link{plot_tshell}()},
\code{\link{write_well_data_to_binary}()}
}
\concept{well}
