% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_non_functional_seqs.R
\name{identify_non_functional_seqs}
\alias{identify_non_functional_seqs}
\title{Identify TCRs that contain non-functional CDR3 sequences}
\usage{
identify_non_functional_seqs(data)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}
}
\value{
A TIRTLseqDataSet object with modified pseudobulk and paired data frames for each
sample. Each dataframe in the ouptut object has added columns that identify
whether the CDR3 alpha and beta
nucleotide sequences contain any stop codons (*) or frame shifts (_) that would
indicate a non-functional chain.

Column \code{is_functional} is TRUE if neither chain has a stop codon or a frame shift.
The columns \code{has_stop_codon} and \code{has_frameshift} are similar, but specific to
each kind of coding error. Other columns identify if the alpha chain or beta chain has
a stop codon or frameshift, and if it is functional.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been deprecated because its functionality has been added to the load_tirtlseq() function

\code{identify_non_functional_seqs()} identifies CDR3 nucleotide sequences in
the data that contain either stop codons (*) or frame shifts (_) that
would indicate a non-functional protein product.
}
\examples{
# example code


}
\seealso{
\code{\link{TCRdist}()}, \code{\link{plot_num_partners}()}

Other deprecated:
\code{\link[=add_single_chain_data]{add_single_chain_data()}},
\code{\link[=identify_paired]{identify_paired()}},
\code{\link[=process_TIRTLseq]{process_TIRTLseq()}}
}
\concept{deprecated}
