% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_paired.R
\name{identify_paired}
\alias{identify_paired}
\title{Identify single chains that are paired in the pseudo-bulk data}
\usage{
identify_paired(data, verbose = TRUE)
}
\arguments{
\item{data}{a TIRTLseq dataset created by \code{\link{load_tirtlseq}()}}
}
\value{
A dataset similar to that created by \code{\link{load_tirtlseq}()}, but
with added columns \code{is_paired} and \code{n_paired} in the alpha and beta
pseudobulk data frames.

\code{is_paired} is TRUE if the chain is found in the
paired data. \code{n_paired} is the number of distinct chains that the particular
chain is paired with.
}
\description{
For each sample in the dataset, \code{identify_paired()} annotates the alpha
and beta pseudobulk data with the number of distinct pairs each chain is a part of
in the paired data as well as a TRUE/FALSE column indicating whether the chain
is paired with any partners.
}
\examples{
# example code
# paired = load_tirtlseq("path_to/your_directory", sep = "_", meta_columns = c("cell_type", "timepoint"))
# paired = identify_paired(paired)

}
\seealso{
\code{\link{load_tirtlseq}()}
}
