% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_env.R
\name{install_python_env}
\alias{install_python_env}
\title{Install python environment}
\usage{
install_python_env()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function install a python environment via the \code{basilisk} package that is
needed for some functions: \code{\link{TCRdist}()}, \code{\link{cluster_tcrs}()}, and \code{\link{run_pairing}()}.

It will run automatically when needed, but is available to the user if needed.
}
\concept{python-deps}
