% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_well_counts_binary.R
\name{load_well_counts_binary}
\alias{load_well_counts_binary}
\title{Load well-level data written to binary format}
\usage{
load_well_counts_binary(folder, prefix, lazy = TRUE)
}
\arguments{
\item{folder}{the directory with ".parquet" and ".rds" files written by \code{\link{write_well_data_to_binary}()}}

\item{prefix}{a prefix with the sample name that will be prepended to the output file names}

\item{lazy}{whether to "lazy load" the clone metadata files. This can speed up
loading because these files are usually very large (default is TRUE).}
}
\value{
This function returns a list with the following elements:
\itemize{
\item \code{alpha} - a sparse matrix ("dgCMatrix") of TCRalpha read counts in column-oriented format.
See \code{\link{Matrix::`dgCMatrix-class`}()}
\item \code{beta} - a sparse matrix ("dgCMatrix") of TCRbeta read counts in column-oriented format.
See \code{\link{Matrix::`dgCMatrix-class`}()}
\item \code{alpha_meta} - a data frame of clone metadata for TCRalpha clones. Row i relates to column i of \code{alpha}.
\item \code{beta_meta} - a data frame of clone metadata for TCRbeta clones. Row i relates to column i of \code{beta}.
\item \code{col_meta} - a data frame of well metadata (i.e. well names)
\item \code{alpha_meta_file} - the file for the TCRalpha clone metadata
\item \code{beta_meta_file} - the file for the TCRbeta clone metadata
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function loads well-level data written to binary files (.parquet and .rds)
by the \code{\link{write_well_data_to_binary}()} function or the
\code{\link{run_pairing}()} function. The data contains
sparse matrices (well x clone) of read counts for TCRalpha and TCRbeta
along with metadata data frames for each clone.

Data loaded this way can be used as input to the \code{\link{plot_tshell}()} function.
}
\seealso{
Other well: 
\code{\link{choose_pair_manual}()},
\code{\link{get_well_subset}()},
\code{\link{get_wells_from_edges}()},
\code{\link{plot_tshell}()},
\code{\link{write_well_data_to_binary}()}
}
\concept{well}
