% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_well_data.R
\name{load_well_data}
\alias{load_well_data}
\title{Load well-level TIRTLseq data}
\usage{
load_well_data(
  folder_path,
  wells = get_well_subset(1:16, 1:24),
  well_pos = 3,
  chain = c("both", "alpha", "beta"),
  nproc = 1L,
  columns = NULL,
  max_files = Inf
)
}
\description{
Load well-level TIRTLseq data
}
\seealso{
Other well: 
\code{\link{filter_well_data}()},
\code{\link{get_well_subset}()},
\code{\link{get_wells_from_edges}()}
}
\concept{well}
