% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_well_data_to_sparse.R
\name{load_well_data_to_sparse_multi}
\alias{load_well_data_to_sparse_multi}
\title{Load well-level TIRTLseq data to sparse matrices}
\usage{
load_well_data_to_sparse_multi(
  folder_path_list,
  sample_names_list,
  wells_list,
  well_pos_list,
  chain = c("alpha", "beta"),
  nproc = 1L,
  columns = NULL,
  max_files = Inf
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other well: 
\code{\link{filter_well_data}()},
\code{\link{get_well_subset}()},
\code{\link{get_wells_from_edges}()},
\code{\link{load_well_data}()}
}
\concept{well}
