% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tcr_schema.R
\name{make_tcr_schema}
\alias{make_tcr_schema}
\title{Make a "schema" for defining a T-Cell Receptor}
\usage{
make_tcr_schema(
  features = c("v", "j", "cdr3_aa", "cdr3_nt"),
  second_alpha = FALSE
)
}
\arguments{
\item{features}{some subset of \code{c("v", "j", "cdr3_aa", "cdr3_nt")} that
you would like to use to define a unique receptor.}

\item{second_alpha}{whether to include a second alpha chain (default is FALSE)}
}
\value{
A vector containing
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is a convenience function that simply returns a vector of column names
that are used by \code{\link{read_external_paired}()} to define a unique
T-Cell receptor.

For example, using \code{features = c("v", "cdr3_aa")} will group all chains
with the same V-alpha/beta and CDR3-alpha/beta amino acid sequence together, while
using \code{features = c("v", "cdr3_aa", "cdr3_nt")} will require that their
nucleotide sequence is also matching.
}
\examples{
make_tcr_schema() ## default: v, j, and cdr3 amino acid and nucleotide sequence
make_tcr_schema(c("v", "j")) ## only v and j segments

}
\seealso{
Other data_processing:
\code{\link[=clean_pairs]{clean_pairs()}},
\code{\link[=combine_bulk_and_paired_data]{combine_bulk_and_paired_data()}},
\code{\link[=filter_duplicate_tcrs]{filter_duplicate_tcrs()}},
\code{\link[=filter_mait]{filter_mait()}},
\code{\link[=filter_nonfunctional_TCRs]{filter_nonfunctional_TCRs()}},
\code{\link[=filter_short_cdr3s]{filter_short_cdr3s()}},
\code{\link[=filter_v_alleles]{filter_v_alleles()}},
\code{\link[=prep_for_tcrdist]{prep_for_tcrdist()}},
\code{\link[=remove_duplicates]{remove_duplicates()}}
}
\concept{data_processing}
