% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clone_size_across_samples.R
\name{plot_clone_size_across_samples}
\alias{plot_clone_size_across_samples}
\title{Line plot of clone read fraction across multiple samples}
\usage{
plot_clone_size_across_samples(
  data,
  clones,
  chain = c("beta", "alpha"),
  pseudo = 1e-06,
  group_vec = NULL,
  sum_readFraction = TRUE,
  samples = NULL,
  return_data = FALSE,
  label_zero = FALSE,
  show_legend = TRUE
)
}
\description{
Line plot of clone read fraction across multiple samples
}
\seealso{
Other plotting: 
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
