% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clonotype_indices.R
\name{plot_clonotype_indices}
\alias{plot_clonotype_indices}
\title{Stacked bar chart with fractions of reads attributed to the most frequent clonotypes}
\usage{
plot_clonotype_indices(
  data,
  chain = c("beta", "alpha"),
  cutoffs = 10^(1:5),
  group_col = NULL,
  label_col = "Sample",
  flip = FALSE,
  return_data = FALSE,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{chain}{the TCR chain to use (default is "beta")}

\item{cutoffs}{the indices used for the end of each group in the bar chart.
The default is 10^(1:5), i.e. the 10th most-frequent clone, the 100th most-frequent clone,
the 1,000th, the 10,000th, and the 100,000th.}

\item{group_col}{(optional) if supplied, a column of the metadata that will be used
to group samples}

\item{label_col}{(optional) labels for the samples}

\item{flip}{(optional) if TRUE, flip the x and y-axes (default is FALSE)}

\item{return_data}{(optional) if TRUE, return the data used to make the plot (default is FALSE)}

\item{color_scheme}{(optional) a color scheme to use}
}
\value{
A list with two objects:

\code{$plot} - a stacked bar chart of relative frequencies of most-frequent clonotypes

\code{$data} - if return_data is TRUE, the data frame used to create the plot
}
\description{
\code{plot_clonotype_indices()} creates a stacked bar chart containing the fraction of
reads for the top 10 most frequent clonotypes, the 11th to 100th most frequent
clonotypes, the 101st to 1000th most frequent, and so on, for each sample in the dataset.
}
\examples{
# example code


}
\seealso{
\code{\link{diversity}()}, \code{\link{plot_diversity}()}

Other plotting: 
\code{\link{plot_clone_size_across_samples}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
