% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_paired_step.R
\name{plot_n_paired_step}
\alias{plot_n_paired_step}
\title{Make a step plot showing the extent of multi-pairing in αβ-TCRs}
\usage{
plot_n_paired_step(
  data,
  chain = c("beta", "alpha", "both"),
  cumulative = TRUE,
  freq = FALSE,
  step = c("vh", "hv"),
  max_x = Inf
)
}
\arguments{
\item{data}{The dataset to use for plotting. Either a "TIRTLseqDataSet" containing
multiple samples or a "TIRTLseqData" object for one sample.}

\item{chain}{The chain to plot: "alpha" or "beta" or "both" (default is "beta")}

\item{cumulative}{Whether to plot the cumulative number of pairs (default is TRUE)}

\item{freq}{Whether to plot the frequency, as opposed to the fraction (default is FALSE)}

\item{step}{The direction of stairs (to pass to \code{geom_step}):  'vh' for vertical then
horizontal, 'hv' for horizontal then vertical.}

\item{max_x}{The maximum x-axis value (number of partners) allowable on the plot (default is Inf).}
}
\value{
A step plot (ggplot object) as defined above.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

For this function, each αβ-TCR pair in the paired data frame is categorized by the
number of partners that one of the chains has (default "beta" chain). Then the fraction
of total pairs is plotted (y-axis) vs. the maximum number of partners allowed (x-axis).

If \code{cumulative = FALSE} the fraction of pairs is plotted for each "number of partners"
(i.e. a probability mass function rather than a cumulative mass function).

If \code{freq = TRUE}, the number of pairs rather than the fraction is plotted.
}
\examples{
load_example_data(dataset = "SJTRC_longitudinal")

plot_n_paired_step(SJTRC_longitudinal)
plot_n_paired_step(SJTRC_longitudinal, cumulative = FALSE)

plot_n_paired_step(SJTRC_longitudinal, freq = TRUE)

plot_n_paired_step(SJTRC_longitudinal, chain = "alpha")
plot_n_paired_step(SJTRC_longitudinal, chain = "both")

}
\seealso{
Other qc:
\code{\link[=get_all_tcrs]{get_all_tcrs()}},
\code{\link[=get_pair_stats]{get_pair_stats()}},
\code{\link[=get_paired_by_read_fraction_range]{get_paired_by_read_fraction_range()}},
\code{\link[=plot_n_reads]{plot_n_reads()}},
\code{\link[=plot_num_partners]{plot_num_partners()}},
\code{\link[=plot_paired]{plot_paired()}},
\code{\link[=plot_paired_by_read_fraction_range]{plot_paired_by_read_fraction_range()}},
\code{\link[=plot_paired_vs_rank]{plot_paired_vs_rank()}},
\code{\link[=plot_pairs_with_eachother]{plot_pairs_with_eachother()}},
\code{\link[=plot_ranks]{plot_ranks()}},
\code{\link[=plot_read_fraction_vs_pair_status]{plot_read_fraction_vs_pair_status()}},
\code{\link[=plot_sample_overlap]{plot_sample_overlap()}},
\code{\link[=summarize_data]{summarize_data()}}
}
\concept{qc}
