% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_reads.R
\name{plot_n_reads}
\alias{plot_n_reads}
\title{Bar plot of the number of single-chain reads for each sample}
\usage{
plot_n_reads(
  data,
  chain = c("both", "beta", "alpha"),
  samples = NULL,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{chain}{the TCR chain to plot (default is both "alpha" and "beta")}

\item{samples}{(optional) which samples to plot (default is all samples)}

\item{color_scheme}{a color scheme for the plot}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function plots the total number of alpha and/or beta reads (default is both)
for each sample.
}
\examples{
load_example_data(dataset = "SJTRC_longitudinal")

plot_n_reads(SJTRC_longitudinal)


}
\seealso{
Other qc:
\code{\link[=get_all_tcrs]{get_all_tcrs()}},
\code{\link[=get_pair_stats]{get_pair_stats()}},
\code{\link[=get_paired_by_read_fraction_range]{get_paired_by_read_fraction_range()}},
\code{\link[=plot_n_paired_step]{plot_n_paired_step()}},
\code{\link[=plot_num_partners]{plot_num_partners()}},
\code{\link[=plot_paired]{plot_paired()}},
\code{\link[=plot_paired_by_read_fraction_range]{plot_paired_by_read_fraction_range()}},
\code{\link[=plot_paired_vs_rank]{plot_paired_vs_rank()}},
\code{\link[=plot_pairs_with_eachother]{plot_pairs_with_eachother()}},
\code{\link[=plot_ranks]{plot_ranks()}},
\code{\link[=plot_read_fraction_vs_pair_status]{plot_read_fraction_vs_pair_status()}},
\code{\link[=plot_sample_overlap]{plot_sample_overlap()}},
\code{\link[=summarize_data]{summarize_data()}}
}
\concept{qc}
