% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranks.R, R/rank_plot.R
\name{plot_ranks}
\alias{plot_ranks}
\alias{rank_plot}
\title{Line plot of clonotype rank vs. read fraction for each sample}
\usage{
plot_ranks(
  data,
  chain = c("beta", "alpha"),
  column = "readFraction",
  color_scheme = NULL
)

rank_plot(...)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{chain}{the chain to plot, "alpha" or "beta" (default is "beta")}

\item{column}{the variable to plot (default is "readFraction")}

\item{color_scheme}{(optional) the color scheme for the plot}
}
\value{
a ggplot object with the rank plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{plot_ranks()} returns a line plot of TCR rank vs. read fraction for a given sample.
}
\examples{
load_example_data(dataset = "SJTRC_longitudinal")

plot_ranks(SJTRC_longitudinal, chain = "beta")


}
\seealso{
Other qc:
\code{\link[=get_all_tcrs]{get_all_tcrs()}},
\code{\link[=get_pair_stats]{get_pair_stats()}},
\code{\link[=get_paired_by_read_fraction_range]{get_paired_by_read_fraction_range()}},
\code{\link[=plot_n_paired_step]{plot_n_paired_step()}},
\code{\link[=plot_n_reads]{plot_n_reads()}},
\code{\link[=plot_num_partners]{plot_num_partners()}},
\code{\link[=plot_paired]{plot_paired()}},
\code{\link[=plot_paired_by_read_fraction_range]{plot_paired_by_read_fraction_range()}},
\code{\link[=plot_paired_vs_rank]{plot_paired_vs_rank()}},
\code{\link[=plot_pairs_with_eachother]{plot_pairs_with_eachother()}},
\code{\link[=plot_read_fraction_vs_pair_status]{plot_read_fraction_vs_pair_status()}},
\code{\link[=plot_sample_overlap]{plot_sample_overlap()}},
\code{\link[=qc_report]{qc_report()}},
\code{\link[=summarize_data]{summarize_data()}}
}
\concept{qc}
