% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_fraction_vs_pair_status.R
\name{plot_read_fraction_vs_pair_status}
\alias{plot_read_fraction_vs_pair_status}
\title{A point plot of read fraction vs. rank for the most frequent alpha/beta chains}
\usage{
plot_read_fraction_vs_pair_status(
  data,
  sample,
  chain = c("both", "beta", "alpha"),
  n_max = 100,
  show_num_partners = F,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{sample}{the sample to plot (either by number or sample id)}

\item{chain}{the chain to plot, "alpha", "beta", or "both" (default is "both")}

\item{n_max}{the number of most frequent single-chains to plot}

\item{show_num_partners}{whether to show the number of partners for each single-chain (default is FALSE)}

\item{color_scheme}{(optional) the color scheme for the plot}
}
\value{
A ggplot object with the point plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function creates two point plots (alpha left, beta right) of the \code{n_max} most frequent
single-chains found in a sample.

This function creates a plot similar to \code{\link{plot_pairs_with_eachother}()},
but does not mirror the alpha chains plot and add connections between paired alpha and beta chains.
}
\examples{
load_example_data(dataset = "SJTRC_longitudinal")

plot_read_fraction_vs_pair_status(SJTRC_longitudinal, sample = 1, n_max = 100)

plot_read_fraction_vs_pair_status(SJTRC_longitudinal, sample = 1, n_max = 100,
show_num_partners = TRUE)

}
\seealso{
Other qc:
\code{\link[=get_all_tcrs]{get_all_tcrs()}},
\code{\link[=get_pair_stats]{get_pair_stats()}},
\code{\link[=get_paired_by_read_fraction_range]{get_paired_by_read_fraction_range()}},
\code{\link[=plot_n_paired_step]{plot_n_paired_step()}},
\code{\link[=plot_n_reads]{plot_n_reads()}},
\code{\link[=plot_num_partners]{plot_num_partners()}},
\code{\link[=plot_paired]{plot_paired()}},
\code{\link[=plot_paired_by_read_fraction_range]{plot_paired_by_read_fraction_range()}},
\code{\link[=plot_paired_vs_rank]{plot_paired_vs_rank()}},
\code{\link[=plot_pairs_with_eachother]{plot_pairs_with_eachother()}},
\code{\link[=plot_ranks]{plot_ranks()}},
\code{\link[=plot_sample_overlap]{plot_sample_overlap()}},
\code{\link[=qc_report]{qc_report()}},
\code{\link[=summarize_data]{summarize_data()}}
}
\concept{qc}
