% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scTIRTLseq_plate.R
\name{plot_scTIRTLseq_plate}
\alias{plot_scTIRTLseq_plate}
\title{Plot a plate of single-cell TIRTL-seq data}
\usage{
plot_scTIRTLseq_plate(df, title = "")
}
\arguments{
\item{df}{a data frame, output from the \code{\link{process_scTIRTLseq}()} function}

\item{title}{an optional title for the plot}
}
\value{
A ggplot object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function plots which wells in a plate are missing an alpha and/or beta chain
for single-cell TIRTL-seq data.
}
\seealso{
Other single-cell:
\code{\link[=process_scTIRTLseq]{process_scTIRTLseq()}}
}
\concept{single-cell}
