% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scTIRTLseq.R
\name{process_scTIRTLseq}
\alias{process_scTIRTLseq}
\title{Process single-cell TIRTL-seq data}
\usage{
process_scTIRTLseq(
  folder,
  min_read_fraction = 0.1,
  min_read_count = 15,
  require_functional = TRUE
)
}
\arguments{
\item{folder}{the path of the folder with .tsv (or .tsv.gz) output files from MiXCR}

\item{min_read_fraction}{the minimum read fraction required to select a chain for a well}

\item{min_read_count}{the minimum read count required to select a chain for a well}

\item{require_functional}{whether to require a functional chain (i.e. without frameshift or stop codons in the CDR3 region)}
}
\value{
A data frame with one TCR pair selected per well.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function reads a directory of MiXCR output files from single-cell TIRTL-seq
and outputs a data frame with a single TCR for each well.
}
\seealso{
Other single-cell:
\code{\link[=plot_scTIRTLseq_plate]{plot_scTIRTLseq_plate()}}
}
\concept{single-cell}
