% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIRTL_process.R, R/process_TIRTLseq.R
\name{TIRTL_process}
\alias{TIRTL_process}
\alias{process_TIRTLseq}
\title{Run data processing functions on a TIRTLseqDataSet object}
\usage{
TIRTL_process(...)

process_TIRTLseq(data, clean = FALSE, remove_nonfunctional = FALSE)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{clean}{(optional) a TRUE/FALSE value, whether or not to "clean" the paired data by removing
excess pairs for individual alpha and beta chains (default is FALSE).}

\item{remove_nonfunctional}{whether to remove non-functional TCR chains (default is FALSE)}
}
\value{
a TIRTLseqDataSet object with annotated and (optionally) cleaned data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function runs annotation and data cleaning functions on a
TIRTLseqDataSet object. Specifically, it calls the functions \code{add_single_chain_data()},
\code{identify_paired()}, \code{identify_non_functional_seqs()}, and (optionally) \code{clean_pairs()}.
}
\seealso{
Other data_processing:
\code{\link[=add_single_chain_data]{add_single_chain_data()}},
\code{\link[=clean_pairs]{clean_pairs()}},
\code{\link[=combine_bulk_and_paired_data]{combine_bulk_and_paired_data()}},
\code{\link[=filter_duplicate_tcrs]{filter_duplicate_tcrs()}},
\code{\link[=filter_mait]{filter_mait()}},
\code{\link[=filter_nonfunctional_TCRs]{filter_nonfunctional_TCRs()}},
\code{\link[=filter_short_cdr3s]{filter_short_cdr3s()}},
\code{\link[=filter_v_alleles]{filter_v_alleles()}},
\code{\link[=identify_non_functional_seqs]{identify_non_functional_seqs()}},
\code{\link[=identify_paired]{identify_paired()}},
\code{\link[=make_tcr_schema]{make_tcr_schema()}},
\code{\link[=prep_for_tcrdist]{prep_for_tcrdist()}},
\code{\link[=read_external_bulk]{read_external_bulk()}},
\code{\link[=read_external_paired]{read_external_paired()}},
\code{\link[=remove_duplicates]{remove_duplicates()}}
}
\concept{data_processing}
