% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranks.R
\name{rank_plot}
\alias{rank_plot}
\title{Line plot of TCR rank vs. frequency}
\usage{
rank_plot(
  data,
  chain = c("alpha", "beta"),
  column = "readFraction",
  color_scheme = NULL
)
}
\arguments{
\item{param1}{the first parameter}

\item{param2}{the second parameter}
}
\value{
A line plot (ggplot object) of TCR rank vs. read fraction
}
\description{
\code{plot_ranks()} returns a line plot of TCR rank vs. read fraction for a given sample.
}
\details{
Fill in later.
}
\examples{
# example code


}
