% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_dupes_paired}
\alias{remove_dupes_paired}
\title{Remove duplicates from paired chain data.}
\usage{
remove_dupes_paired(data)
}
\description{
There will be duplicates in paired chain data when a pair is called by both
the T-SHELL and MAD-HYPE algorithms.
The input 'data' needs to be either a single data frame (paired chain) or a list
of data frames (paired chain)
}
