% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_samples.R
\name{reorder_samples}
\alias{reorder_samples}
\title{Re-order samples in a TIRTLseqDataSet object}
\usage{
reorder_samples(data, samples)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{samples}{the samples of 'data', in their desired order.

Either:
\enumerate{
\item A numeric vector of indices
\item A character vector of sample names
\item A character vector of metadata conditions where each entry is of the form \code{"column==value"}
}}
}
\value{
A TIRTLseqDataSet object with re-ordered samples.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function changes the order of the samples in a TIRTLseqDataSet object to the order that the user specifies.
}
\examples{
load_example_data(dataset = "SJTRC_longitudinal")
print(SJTRC_longitudinal$meta)
new_order = names(SJTRC_longitudinal$data) \%>\% rev()
print(new_order)
SJTRC_longitudinal = reorder_samples(SJTRC_longitudinal, new_order)
print(SJTRC_longitudinal$meta)

}
\seealso{
Other data_wrangling:
\code{\link[=add_metadata]{add_metadata()}},
\code{\link[=filter_dataset]{filter_dataset()}}
}
\concept{data_wrangling}
