% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_overlap.R
\name{sample_overlap}
\alias{sample_overlap}
\title{Calculate and plot agreement/overlap between samples}
\usage{
sample_overlap(
  data,
  chain = c("paired", "alpha", "beta"),
  n_seq = 200,
  show_row_names = FALSE,
  show_column_names = FALSE,
  label_col = "Sample",
  title = ""
)
}
\arguments{
\item{data}{the dataset, an object loaded using the \code{\link{load_tirtlseq}()} function}

\item{chain}{which chain to plot: either paired or alpha-/beta-pseudobulk. (default "paired")}

\item{n_seq}{the number of most frequent TCR sequences to use (default 200)}

\item{show_row_names}{whether to show row names for the heatmap (default FALSE)}

\item{show_column_names}{whether to show column names for the heatmap (default FALSE)}

\item{label_col}{a column of the metadata to use as labels for rows and columns (default "Sample", uses the sample_id)}

\item{title}{a title for the heatmap}
}
\value{
A heatmap with hierarchically clustered rows and columns showing the number of
TCRs shared between each pair of samples, among their top N most frequent TCRs.
}
\description{
\code{sample_overlap()} returns a heatmap showing the overlap in most frequent TCRs
among pairs of samples in a dataset
}
\details{
The function takes the top N most frequent TCRs found in each dataset (default 200) and
compares their overlap between samples.
}
\examples{
# example code
# data = load_tirtlseq("your_directory/")
# sample_overlap(data, chain = "paired", n_seq=200)

}
\seealso{
\code{\link{load_tirtlseq}()}
}
