% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{submat}
\alias{submat}
\title{Substitution penalty matrix for TCRdist amino acids and V-segments}
\format{
A symmetric matrix with substitution penalties for pairs of amino acids or
V-segments for TCRdist calculation. The amino acids/V-segments corresponding to each
row/column are listed in the "params" table.
}
\usage{
submat
}
\description{
Substitution penalty matrix for TCRdist amino acids and V-segments
}
\concept{data}
\keyword{datasets}
