% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_scTIRTLseq.R
\name{summarize_scTIRTLseq}
\alias{summarize_scTIRTLseq}
\title{Returns a summary of single-cell TIRTLseq data, printing the total number of
wells with data, number of wells missing an alpha/beta chain, number of wells
with a second alpha chain, etc.}
\usage{
summarize_scTIRTLseq(df)
}
\description{
Returns a summary of single-cell TIRTLseq data, printing the total number of
wells with data, number of wells missing an alpha/beta chain, number of wells
with a second alpha chain, etc.
}
\seealso{
Other single-cell: 
\code{\link{clean_scTIRTLseq}()}
}
\concept{single-cell}
