% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRdist_old.R
\name{TCRdist_old}
\alias{TCRdist_old}
\title{GPU implementation of TCRdist, a distance/similarity metric for pairs of TCRs}
\usage{
TCRdist_old(
  tcr1,
  tcr2 = NULL,
  remove_MAIT = FALSE,
  params = NULL,
  submat = NULL,
  tcrdist_cutoff = NULL,
  chunk_size = 1000,
  print_chunk_size = 10,
  print_res = TRUE,
  only_lower_tri = TRUE,
  return_data = TRUE,
  write_to_tsv = FALSE,
  backend = c("auto", "cpu", "cupy", "mlx"),
  fork = NULL,
  shared = NULL
)
}
\arguments{
\item{tcr1}{a data frame with one TCR per row. It must have the columns "va", "vb", "cdr3a", and "cdr3b"}

\item{tcr2}{(optional) another data frame of TCRs. If supplied, TCRdist will be calculated
for every combination of one TCR from tcr1 and one TCR from tcr2. Otherwise, it will calculate TCRdist
for every pair of TCRs in tcr1.}

\item{remove_MAIT}{whether to remove TCRs from MAIT cells (default is FALSE)}

\item{params}{(optional) a table of valid parameters for amino acids and va/vb segments.
(default is NULL, which uses TIRTLtools::params)}

\item{submat}{(optional) a substitution matrix with mismatch penalties for each
combination of amino acids or va/vb segments (default is NULL, which uses TIRTLtools::submat).}

\item{tcrdist_cutoff}{(optional) discard all TCRdist values above this cutoff. If not supplied by the user, this will default to 90 for dual-chain TCRdist or 45 for single-chain TCRdist.}

\item{chunk_size}{(optional) The chunk size to use in calculation of TCRdist (default 1000). If set at n,
it will calculate pairwise TCRdist for n x n TCRs at once. This may be as high as allowable by GPU memory
(in our testing, a chunk_size of 1000 to 5000 provided the fastest runtime and
chunk_size of over 7500 resulted in memory errors on some GPUs).}

\item{print_chunk_size}{(optional) print a line of output for every n TCRs processed (default 1000)}

\item{print_res}{(optional) print summary of results (default is TRUE)}

\item{only_lower_tri}{(optional) return one TCRdist value for each pair (like the lower triangle of a symmetric matrix). Default is TRUE.}

\item{return_data}{(optional) whether to return the output result from the function.
With large data it may be desirable to write the result to disk instead. (default is TRUE, returns output)}

\item{write_to_tsv}{(optional) write the results to a tab-separated file ".tsv" (default is FALSE, does not write .tsv file)}

\item{backend}{(optional) the CPU or GPU backend to use (default "auto")}

\item{fork}{(optional) a TRUE/FALSE value for whether to "fork" a new Python process for running TCRdist via the "basilisk" package. Default is NULL, which should use choose a safe value based on how the package is loaded.}

\item{shared}{(optional) a TRUE/FALSE value for whether to "share" the Python process for running TCRdist via the "basilisk" package. Default is NULL, which should use choose a safe value based on how the package is loaded.}
}
\value{
A list with entries:

\code{$TCRdist_df} - a data frame with three columns: "node1_0index", "node2_0index", and "TCRdist".
The first two columns contain the indices (0-indexed) of the TCRs for each pair.
The last column contains the TCRdist if it is below the cutoff. The output is sparse in that it only contains
pairs that have TCRdist <= cutoff.

\code{$tcr1} - a data frame of the TCRs supplied to the function. It contains an additional column
"tcr_index" with the (0-indexed) index of each TCR.

\code{$tcr2} - a similar data frame for tcr2, if it was supplied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

An efficient, batched version of TCRdist that is compatible with both NVIDIA and Apple Silicon GPUs.
}
\details{
This function calculates pairwise TCRdist (Dash et al., Nature 2017) for a set of TCRs
(or between two sets of TCRs) and returns a sparse output with the TCRdist and indices of all pairs
that have TCRdist less than or equal to a desired cutoff (default cutoff is 90).

The function uses the \code{reticulate} package to call a python script that uses \code{cupy} (NVIDIA GPUs), \code{mlx} (Apple Silicon GPUs),
or \code{numpy} (no GPU) to calculate TCRdist efficiently.
}
\examples{
load_example_data(dataset = "SJTRC_minimal")
df = get_all_tcrs(SJTRC_minimal, chain="paired", remove_duplicates = TRUE)
result = TCRdist(df, tcrdist_cutoff = 90)
edge_df = result[['TCRdist_df']] \%>\%
  data.table::as.data.table() ### table of TCRdist values <= cutoff
node_df = result[['tcr1']] \%>\%
  data.table::as.data.table() ### table of input data with indices

edge_df ## sparse 3-column output: node1, node2, TCRdist
## note that indices start at 0 and are found in node_df$tcr_index

node_df \%>\%
  select(tcr_index, everything()) \%>\%
  mutate(alpha_nuc = paste(substr(alpha_nuc, 0, 20), "...", sep = ""),
         beta_nuc = paste(substr(beta_nuc, 0, 20), "...", sep = ""))

}
\seealso{
\code{\link{cluster_tcrs}()}, \code{\link{plot_clusters}()}, and \code{\link{identify_non_functional_seqs}()}

Other tcr_similarity:
\code{\link[=TCRdist]{TCRdist()}},
\code{\link[=TCRdist_cpp]{TCRdist_cpp()}},
\code{\link[=TCRdist_to_igraph]{TCRdist_to_igraph()}},
\code{\link[=TCRdist_to_sparse_matrix]{TCRdist_to_sparse_matrix()}},
\code{\link[=add_to_tcr_network]{add_to_tcr_network()}},
\code{\link[=cluster_tcrs]{cluster_tcrs()}},
\code{\link[=plot_clusters]{plot_clusters()}}
}
\concept{tcr_similarity}
\keyword{internal}
