% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIRTL_process.R
\name{TIRTL_process}
\alias{TIRTL_process}
\title{Run data processing functions on a TIRTLseqData object}
\usage{
TIRTL_process(data, clean = FALSE)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{clean}{(optional) a TRUE/FALSE value, whether or not to "clean" the paired data by removing
excess pairs for individual alpha and beta chains (default is FALSE).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function runs annotation and data cleaning functions on a
TIRTLseqData object. Specifically, it calls the functions \code{add_single_chain_data()},
\code{identify_paired()}, \code{identify_non_functional_seqs()}, and (optionally) \code{clean_pairs()}.
}
\seealso{
Other data_processing: 
\code{\link{add_single_chain_data}()},
\code{\link{clean_pairs}()},
\code{\link{filter_mait}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()},
\code{\link{remove_duplicates}()}
}
\concept{data_processing}
