% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to a TIRTLseqData object}
\usage{
add_metadata(obj, meta_columns = NULL, sep = "_")
}
\arguments{
\item{obj}{a TIRTLseqData object}

\item{meta_columns}{(optional) a vector of names for metadata for each field in the sample names.}

\item{sep}{(optional) the character separating fields in the sample names of the data
For example \code{c("marker", "timepoint", "donor")} for samples named something
like "cd8_timepoint2_donor1".}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other data_wrangling: 
\code{\link{filter_dataset}()},
\code{\link{load_tirtlseq}()},
\code{\link{reorder_samples}()}
}
\concept{data_wrangling}
