% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pseudobulk_counts_to_paired_data.R
\name{add_single_chain_data}
\alias{add_single_chain_data}
\title{The input 'data' must be a TIRTL-seq data object loaded with the loading function.
i.e. it must be a list of length two, with data$meta and data$data, where
data$data is a list of samples, each having data for paired, alpha, and beta chains.}
\usage{
add_single_chain_data(data)
}
\description{
The input 'data' must be a TIRTL-seq data object loaded with the loading function.
i.e. it must be a list of length two, with data$meta and data$data, where
data$data is a list of samples, each having data for paired, alpha, and beta chains.
}
