% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_single_chain_data.R
\name{add_single_chain_data}
\alias{add_single_chain_data}
\title{Add single-chain pseudo-bulk counts/proportions to paired TIRTL-seq data}
\usage{
add_single_chain_data(data, verbose = TRUE)
}
\arguments{
\item{data}{a TIRTLseq dataset created by \code{\link{load_tirtlseq}()}}
}
\value{
A TIRTLseq dataset object where the paired data frames for each sample have
added columns for read counts and proportions from the single-chain pseudobulk data.
}
\description{
The \code{add_single_chain_data()} function adds read counts and proportions
from the single-chain pseudobulk data to the paired data frame for each sample
of a dataset.
}
\examples{
# example code


}
\seealso{
\code{\link{load_tirtlseq}()}
}
