% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_single_chain_data.R
\name{add_single_chain_data}
\alias{add_single_chain_data}
\title{Add single-chain read counts/fractions to the paired TCR data}
\usage{
add_single_chain_data(data, verbose = TRUE)
}
\arguments{
\item{data}{a TIRTLseq dataset created by \code{\link{load_tirtlseq}()}}

\item{verbose}{(optional) whether to print progress of the function (default is TRUE).}
}
\value{
A TIRTLseq dataset object where the paired data frames for each sample have
added columns for read counts and proportions from the single-chain pseudobulk data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The \code{add_single_chain_data()} function adds read counts and proportions
from the single-chain pseudobulk data to the paired data frame for each sample
of a dataset.
}
\seealso{
Other data_processing: 
\code{\link{TIRTL_process}()},
\code{\link{clean_pairs}()},
\code{\link{filter_duplicate_tcrs}()},
\code{\link{filter_mait}()},
\code{\link{filter_nonfunctional_TCRs}()},
\code{\link{filter_short_cdr3s}()},
\code{\link{filter_v_alleles}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()},
\code{\link{remove_duplicates}()}
}
\concept{data_processing}
