% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pairs.R
\name{clean_pairs}
\alias{clean_pairs}
\title{Clean paired TCR data, keeping at most one beta-chain partner for each
individual alpha-chain and at most two alpha-chain partners for each individual beta-chain}
\usage{
clean_pairs(data, n_max_alpha = 2, n_max_beta = 1, verbose = TRUE)
}
\description{
Clean paired TCR data, keeping at most one beta-chain partner for each
individual alpha-chain and at most two alpha-chain partners for each individual beta-chain
}
\seealso{
Other data_processing: 
\code{\link{TIRTL_process}()},
\code{\link{add_single_chain_data}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()},
\code{\link{remove_duplicates}()}
}
\concept{data_processing}
