% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_scTIRTLseq.R
\name{clean_scTIRTLseq}
\alias{clean_scTIRTLseq}
\title{Imputes missing alpha and beta chains where possible for single-cell TIRTLseq data}
\usage{
clean_scTIRTLseq(df, verbose = TRUE, keep_all_columns = FALSE)
}
\arguments{
\item{df}{a data frame with single-cell TIRTLseq data}

\item{verbose}{print a before and after summary of clonotypes (default TRUE)}

\item{keep_all_columns}{keep intermediate columns created by the function (default FALSE)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other single-cell: 
\code{\link{summarize_scTIRTLseq}()}
}
\concept{single-cell}
