% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_tcrs.R
\name{cluster_tcrs}
\alias{cluster_tcrs}
\title{Cluster TCRs (using the Leiden algorithm) based on their pairwise TCRdist values}
\usage{
cluster_tcrs(
  data,
  tcrdist_cutoff = NULL,
  resolution = 0.1,
  with_db = TRUE,
  db = TIRTLtools::vdj_db,
  allow_self_edges = TRUE,
  remove_MAIT = TRUE
)
}
\arguments{
\item{data}{a list of TIRTLseq TCR data for samples created with \code{\link{load_tirtlseq}()}}

\item{tcrdist_cutoff}{discard all TCRdist values above this cutoff. If not
supplied by the user, this will default to 90 for dual-chain TCRdist or 45
for single-chain TCRdist. Note: Higher cutoffs will return more
data, at most NxN where N is the number of unique TCRs.}

\item{resolution}{the "resolution" parameter for the Leiden algorithm. A lower
value will produce larger clusters and a higher value will produce smaller clusters.
Typical values are in the 0.1 - 2.0 range. A higher value may be better for densely
connected data while a lower value may be better for moderately connected data. Default is 0.1.}

\item{with_db}{if TRUE, observed clones will be compared and clustered with a dataframe of
annotated clones. By default, a dataframe with VDJ-db annotations is used.}

\item{db}{a data frame with annotated TCRs. The default is the VDJ-db database.}

\item{allow_self_edges}{if FALSE, only calculate TCRdist between input data TCRs
and the TCR annotation database (db). If TRUE, calculate pairwise TCRdist for all
of the data including the input and the annotated TCRs.}

\item{remove_MAIT}{remove MAIT TCRs before clustering (default is TRUE)}
}
\value{
Returns a list with the following elements:

\code{$df} - a data frame with all unique TCRs along with cluster annotations

\code{$dist_df} - a data frame with distances (TCRdist) between TCR pairs in long format

\code{$sparse_adj_mat} - an adjacency matrix (in sparse format) marking TCR pairs with TCRdist <= tcrdist_cutoff

\code{$graph_adj} - an igraph object created from the adjacency matrix

\code{$tcrdist_cutoff} - the cutoff used for TCRdist

\code{$resolution} - the resolution parameter used for the Leiden algorithm
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{cluster_tcrs()} function aggregates all of the paired TCRs found in the data,
calculates pairwise similarity using the va, vb, cdr3a, and cdr3b regions (via TCRdist),
and clusters the results using the Leiden algorithm.
}
\details{
The function also filters the dataset to TCRs that are valid for TCRdist.

The following TCRs are removed:
\itemize{
\item TCRs that contain stop codons (*) or frame shifts (_) in their cdr3a or cdr3b regions
\item TCRs that contain a cdr3 region with 5 or less amino acids
\item TCRs that contain a v segment allele not found in our parameter table
}

V-segments that do not specify an allele (e.g. "TRAV1-2" instead of "TRAV1-2*01")
will be assigned to the "*01" allele.
}
\examples{
load_example_data(dataset = "SJTRC_minimal")
df = get_all_tcrs(SJTRC_minimal, chain="paired", remove_duplicates = TRUE)

result = cluster_tcrs(df)

}
\seealso{
\code{\link{plot_clusters}()}, \code{\link{identify_non_functional_seqs}()}, \code{\link{TCRdist}()}

Other tcr_similarity:
\code{\link[=TCRdist]{TCRdist()}},
\code{\link[=TCRdist_cpp]{TCRdist_cpp()}},
\code{\link[=TCRdist_old]{TCRdist_old()}},
\code{\link[=plot_clusters]{plot_clusters()}}
}
\concept{tcr_similarity}
