% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_bulk_and_paired_data.R
\name{combine_bulk_and_paired_data}
\alias{combine_bulk_and_paired_data}
\title{Combine bulk single-chain TCR data with paired TCR data}
\usage{
combine_bulk_and_paired_data(alpha, beta, paired, meta)
}
\arguments{
\item{alpha}{a dataframe with alpha chain bulk TCR data}

\item{beta}{a dataframe with beta chain bulk TCR data}

\item{paired}{a dataframe with paired chain TCR data}

\item{meta}{either the sample name or a metadata data frame containing the sample name in the 'sample_id' column.}
}
\value{
a pairedTCRDataSet object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function...
}
\seealso{
Other data_processing:
\code{\link[=TIRTL_process]{TIRTL_process()}},
\code{\link[=add_single_chain_data]{add_single_chain_data()}},
\code{\link[=clean_pairs]{clean_pairs()}},
\code{\link[=filter_duplicate_tcrs]{filter_duplicate_tcrs()}},
\code{\link[=filter_mait]{filter_mait()}},
\code{\link[=filter_nonfunctional_TCRs]{filter_nonfunctional_TCRs()}},
\code{\link[=filter_short_cdr3s]{filter_short_cdr3s()}},
\code{\link[=filter_v_alleles]{filter_v_alleles()}},
\code{\link[=identify_non_functional_seqs]{identify_non_functional_seqs()}},
\code{\link[=identify_paired]{identify_paired()}},
\code{\link[=make_tcr_schema]{make_tcr_schema()}},
\code{\link[=prep_for_tcrdist]{prep_for_tcrdist()}},
\code{\link[=read_external_bulk]{read_external_bulk()}},
\code{\link[=read_external_paired]{read_external_paired()}},
\code{\link[=remove_duplicates]{remove_duplicates()}}
}
\concept{data_processing}
