% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download TIRTLseq datasets}
\usage{
download_dataset(dataset = "SJTRC_TIRTLseq_minimal", folder = tempdir())
}
\arguments{
\item{dataset}{The TIRTLseq dataset to download}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\concept{download}
