% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{filter_dataset}
\alias{filter_dataset}
\title{Subset a TIRTLseqData object to only the requested samples}
\usage{
filter_dataset(data, samples)
}
\arguments{
\item{data}{a TIRTLseq dataset created by \code{\link{load_tirtlseq}()}}

\item{samples}{the selected samples. Either1) a numeric vector of indices,
2) a character vector of sample names, or 3) a character vector of metadata
conditions where each entry is of the form \code{"column==value"}.}
}
\value{
A dataset object similar to that created by \code{\link{load_tirtlseq}()},
but with only the selected samples.
}
\description{
The \code{filter_dataset()} function is used to select a subset of samples from
a loaded TIRTLseq dataset and create a new dataset object.
}
\details{
The function accepts 1) a numeric vector of indices, 2) a character vector of sample names, or
3) a character vector of metadata conditions where each entry is of the form \code{"column==value"}.

In the third case, \code{c("cell_type==cd4", "timepoint==tp2")} would, for example, select
samples whose \code{cell_type} is \code{cd4} and whose \code{timepoint} is \code{tp2} in
the sample metadata.
}
\examples{
# example code
# paired = load_tirtlseq("path_to/your_directory", sep = "_", meta_columns = c("cell_type", "timepoint"))
# p2 = filter_dataset(paired, 1:3) ### by indices
# p3 = filter_dataset(paired, c("cd8_tp1_v2", "cd8_tp2_v2", "cd8_tp3_v2")) ### by sample names
# p4 = filter_dataset(paired, "cell_type==cd4") ### by sample metadata condition
# p5 = filter_dataset(paired, c("cell_type==cd4", "timepoint==tp2")) ### by multiple sample metadata conditions

}
\seealso{
Other data_wrangling: 
\code{\link{add_metadata}()},
\code{\link{load_tirtlseq}()},
\code{\link{reorder_samples}()}
}
\concept{data_wrangling}
