% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicate_tcrs.R
\name{filter_duplicate_tcrs}
\alias{filter_duplicate_tcrs}
\title{Remove duplicate TCRs from a data frame}
\usage{
filter_duplicate_tcrs(df, verbose = TRUE)
}
\arguments{
\item{df}{data frame with paired TCRs}

\item{verbose}{whether to print number of TCRs removed}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
By default, the TIRTLseq pairing scripts write a "paired" data table that has one row
per pair per pairing algorithm (MAD-HYPE or T-SHELL). Thus, any pairs identified by both methods
will be found in two rows. This function removes those duplicates.
}
\seealso{
Other data_processing: 
\code{\link{TIRTL_process}()},
\code{\link{add_single_chain_data}()},
\code{\link{clean_pairs}()},
\code{\link{filter_mait}()},
\code{\link{filter_nonfunctional_TCRs}()},
\code{\link{filter_short_cdr3s}()},
\code{\link{filter_v_alleles}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()},
\code{\link{remove_duplicates}()}
}
\concept{data_processing}
