% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mait.R
\name{filter_mait}
\alias{filter_mait}
\title{Remove MAIT (Mucosal-Associated Invariant T cells) TCRs}
\usage{
filter_mait(df, verbose = TRUE)
}
\arguments{
\item{df}{data frame with paired TCRs}

\item{verbose}{whether to print number of MAIT TCRs removed}
}
\description{
Remove MAIT (Mucosal-Associated Invariant T cells) TCRs
}
\seealso{
Other data_processing: 
\code{\link{TIRTL_process}()},
\code{\link{add_single_chain_data}()},
\code{\link{clean_pairs}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()},
\code{\link{remove_duplicates}()}
}
\concept{data_processing}
