% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nonfunctional_TCRs.R
\name{filter_nonfunctional_TCRs}
\alias{filter_nonfunctional_TCRs}
\title{Remove TCRs with nonfunctional CDR3 amino acid sequences}
\usage{
filter_nonfunctional_TCRs(data, verbose = TRUE)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{verbose}{whether to print number of TCRs removed}
}
\value{
Returns a TIRTLseqDataSet object with nonfunctional TCRs removed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function removes TCR pairs where the CDR3-alpha or beta segment does not make a
functional protein, i.e. amino acid sequence contains a stop codon (*) or a frameshift (_).
}
\seealso{
Other data_processing: 
\code{\link{TIRTL_process}()},
\code{\link{add_single_chain_data}()},
\code{\link{clean_pairs}()},
\code{\link{filter_duplicate_tcrs}()},
\code{\link{filter_mait}()},
\code{\link{filter_short_cdr3s}()},
\code{\link{filter_v_alleles}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()},
\code{\link{remove_duplicates}()}
}
\concept{data_processing}
