% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_short_cdr3s.R
\name{filter_short_cdr3s}
\alias{filter_short_cdr3s}
\title{Remove TCRs with short CDR3 loops}
\usage{
filter_short_cdr3s(df, verbose = TRUE, min_aa = 6)
}
\arguments{
\item{df}{data frame with paired TCRs}

\item{verbose}{whether to print number of TCRs removed}

\item{min_aa}{the minimum number of amino acids allowed for a CDR3 sequence}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function removes TCR pairs where the CDR3 amino acid segment is short.
By defualt, it removes TCRs where CDR3-alpha or CDR3-beta is less than 6 amino acids long.
}
\seealso{
Other data_processing: 
\code{\link{TIRTL_process}()},
\code{\link{add_single_chain_data}()},
\code{\link{clean_pairs}()},
\code{\link{filter_duplicate_tcrs}()},
\code{\link{filter_mait}()},
\code{\link{filter_nonfunctional_TCRs}()},
\code{\link{filter_v_alleles}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()},
\code{\link{remove_duplicates}()}
}
\concept{data_processing}
