% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_v_alleles.R
\name{filter_v_alleles}
\alias{filter_v_alleles}
\title{Remove TCRs with unknown V-segments}
\usage{
filter_v_alleles(df, params = NULL, verbose = TRUE)
}
\arguments{
\item{df}{data frame with paired TCRs}

\item{params}{a table with acceptable V-segments (default is \code{TIRTLtools::params})}

\item{verbose}{whether to print number of TCRs removed}
}
\value{
a paired TCR data frame with TCRs with unknown V-segments removed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function removes TCR pairs with V-segments that are not found in \code{\link{TIRTLtools::params}()}.
This is needed for pre-processing for \code{\link{TCRdist}()} and is part of \code{\link{prep_for_tcrdist}()},
which is automatically run during \code{\link{TCRdist}()}.
}
\seealso{
Other data_processing: 
\code{\link{TIRTL_process}()},
\code{\link{add_single_chain_data}()},
\code{\link{clean_pairs}()},
\code{\link{filter_duplicate_tcrs}()},
\code{\link{filter_mait}()},
\code{\link{filter_nonfunctional_TCRs}()},
\code{\link{filter_short_cdr3s}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()},
\code{\link{remove_duplicates}()}
}
\concept{data_processing}
