% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_div_metrics.R
\name{get_all_div_metrics}
\alias{get_all_div_metrics}
\title{Returns all diversity metric options for \code{\link{calculate_diversity}()}}
\usage{
get_all_div_metrics()
}
\value{
A vector of all available diversity metrics for the \code{\link{calculate_diversity}()} function.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# example code
get_all_div_metrics()

}
\seealso{
\code{\link{diversity}()}, \code{\link{plot_diversity}()}

Other repertoire_analysis: 
\code{\link{TCRdist}()},
\code{\link{TCRdist_cpp}()},
\code{\link{cluster_tcrs}()},
\code{\link{get_all_tcrs}()},
\code{\link{get_pair_stats}()},
\code{\link{get_paired_by_read_fraction_range}()},
\code{\link{summarize_data}()}
}
\concept{repertoire_analysis}
