% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_tcrs.R
\name{get_all_tcrs}
\alias{get_all_tcrs}
\title{Returns all of the paired TCRs from all samples in a dataset}
\usage{
get_all_tcrs(
  data,
  chain = c("paired", "alpha", "beta"),
  remove_duplicates = TRUE
)
}
\arguments{
\item{data}{a TIRTLseqData object created by \code{\link{load_tirtlseq}()}}

\item{chain}{the TCR chain, "alpha", "beta", or "paired" (default is paired)}

\item{remove_duplicates}{only return one TCR for TCRs paired by both
the T-SHELL and MAD-HYPE algorithms (default is TRUE).}
}
\value{
A dataframe including all of the TCRs in a dataset.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The \code{get_all_tcrs()} function aggregates the TCRs from all samples of a dataset
and puts them into one table.
}
\details{
A pair of TCRs is included twice in the TIRTLseq data if it is recognized by both
the T-SHELL and MAD-HYPE algorithms. If remove_duplicates is TRUE (default) the
function will only return one of these pairs of TCRs.
}
\examples{
# example code


}
\seealso{
\code{\link{load_tirtlseq}()}

Other repertoire_analysis: 
\code{\link{TCRdist}()},
\code{\link{TCRdist_cpp}()},
\code{\link{cluster_tcrs}()},
\code{\link{get_all_div_metrics}()},
\code{\link{get_pair_stats}()},
\code{\link{get_paired_by_read_fraction_range}()},
\code{\link{summarize_data}()}
}
\concept{repertoire_analysis}
