% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_tcrs.R
\name{get_all_tcrs}
\alias{get_all_tcrs}
\title{Aggregate all of the T-cell receptors from a dataset}
\usage{
get_all_tcrs(
  data,
  chain = c("paired", "alpha", "beta"),
  remove_duplicates = TRUE
)
}
\arguments{
\item{data}{a TIRTLseq dataset created by \code{\link{load_tirtlseq()}}}

\item{chain}{the TCR chain, "alpha", "beta", or "paired" (default is paired)}

\item{remove_duplicates}{only return one TCR pair for TCRs recognized by both
the T-SHELL and MAD-HYPE algorithms (default is TRUE).}
}
\value{
A dataframe including all of the TCRs in a dataset.
}
\description{
The \code{get_all_tcrs()} function aggregates the TCRs from all samples of a dataset
and puts them into one table.
}
\details{
A pair of TCRs is included twice in the TIRTLseq data if it is recognized by both
the T-SHELL and MAD-HYPE algorithms. If remove_duplicates is TRUE (default) the
function will only return one of these pairs of TCRs.
}
\examples{
# example code


}
\seealso{
\code{\link{load_tirtlseq()}}
}
