% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pair_stats.R
\name{get_pair_stats}
\alias{get_pair_stats}
\title{Count the number of single-chains that were paired by each algorithm}
\usage{
get_pair_stats(data, verbose = TRUE)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{verbose}{(optional) whether to print progress of the function (default is TRUE).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other repertoire_analysis: 
\code{\link{TCRdist}()},
\code{\link{TCRdist_cpp}()},
\code{\link{cluster_tcrs}()},
\code{\link{get_all_div_metrics}()},
\code{\link{get_all_tcrs}()},
\code{\link{get_paired_by_read_fraction_range}()},
\code{\link{summarize_data}()}
}
\concept{repertoire_analysis}
