% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paired_by_read_fraction_range.R
\name{get_paired_by_read_fraction_range}
\alias{get_paired_by_read_fraction_range}
\title{Calculate the number and fraction of single chains that were paired within
different read fraction ranges}
\usage{
get_paired_by_read_fraction_range(
  data,
  chain = c("beta", "alpha"),
  cutoffs = 10^(-6:-1)
)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{chain}{the TCR chain to use for read fraction (default is "beta")}

\item{cutoffs}{a vector of cutoffs for the read fraction ranges}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other repertoire_analysis: 
\code{\link{TCRdist}()},
\code{\link{TCRdist_cpp}()},
\code{\link{cluster_tcrs}()},
\code{\link{get_all_div_metrics}()},
\code{\link{get_all_tcrs}()},
\code{\link{get_pair_stats}()},
\code{\link{summarize_data}()}
}
\concept{repertoire_analysis}
