% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tshell_settings.R
\name{get_tshell_settings}
\alias{get_tshell_settings}
\title{Get p-value and well threshold settings for T-SHELL}
\usage{
get_tshell_settings(format = c("auto", "384_well", "96_well"))
}
\arguments{
\item{format}{a selection from "auto", "384_well", and "96_well".}
}
\value{
A named list with slots "pval_thres_tshell" and "wij_thres_tshell". If format is
"384_well", the default p-value threshold for a TCR pair is <1e-10 and threshold
for number of wells with both chains is >2 wells. If format is "96_well", these are
<1e-3 and >3 wells respectively. If format is "auto", the selection will be made
based on the number of wells passing QC in the sample. If >= 150 wells pass QC,
the "384_well" settings will be used. Otherwise the "96_well" settings will be used.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function returns a list with p-value and well threshold settings for
384-well or 96-well T-SHELL.
}
\examples{
get_tshell_settings("96_well")
get_tshell_settings("384_well")


}
\seealso{
Other pairing:
\code{\link[=run_pairing]{run_pairing()}}
}
\concept{pairing}
