% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_well_subset.R
\name{get_well_subset}
\alias{get_well_subset}
\title{Get well names from numerical rows and columns}
\usage{
get_well_subset(row_range = 1:16, col_range = 1:24)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other well: 
\code{\link{filter_well_data}()},
\code{\link{get_wells_from_edges}()},
\code{\link{load_well_data}()},
\code{\link{load_well_data_to_sparse_multi}()}
}
\concept{well}
