% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wells_from_edges.R
\name{get_wells_from_edges}
\alias{get_wells_from_edges}
\title{Get well names from top left and bottom right wells}
\usage{
get_wells_from_edges(
  top_left,
  bottom_right,
  return_type = c("wells", "rows_and_columns")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other well: 
\code{\link{filter_well_data}()},
\code{\link{get_well_subset}()},
\code{\link{load_well_data}()}
}
\concept{well}
