% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_non_functional_seqs.R
\name{identify_non_functional_seqs}
\alias{identify_non_functional_seqs}
\title{Identify non-functional CDR3 nucleotide sequences}
\usage{
identify_non_functional_seqs(df)
}
\arguments{
\item{df}{a dataframe of TIRTLseq data, either paired or single-chain pseudobulk for one sample}
}
\value{
A dataframe with added columns that identify whether the CDR3 alpha and beta
nucleotide sequences contain any stop codons (*) or frame shifts (_) that would
indicate a non-functional chain.

If column \code{is_functional} is TRUE if neither chain has a stop codon or a frame shift.
The columns \code{has_stop_codon} and \code{has_frameshift} are similar, but specific to
each kind of coding error. Other columns identify if the alpha chain or beta chain has
a stop codon or frameshift, and if it is functional.
}
\description{
\code{identify_non_functional_seqs()} identifies CDR3 nucleotide sequences in
the data that contain errors, either stop codons (*) or frame shifts (_) that
would indicate a non-functional chain.
}
\examples{
# example code


}
\seealso{
\code{\link{TCRdist}()}, \code{\link{plot_num_partners}()}
}
