% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_example_data}
\alias{load_example_data}
\title{Load example TIRTL-seq data}
\usage{
load_example_data(
  dataset = c("SJTRC_minimal", "SJTRC_longitudinal"),
  verbose = TRUE
)
}
\arguments{
\item{dataset}{either "SJTRC_minimal" or "SJTRC_longitudinal".}

\item{verbose}{whether to print messages to the user (default TRUE)}
}
\value{
The function returns NULL invisibly, but adds a
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This is a helper function to load example TIRTL-seq data.
The function currently loads one of two available TIRTL-seq datasets. It saves the data
to the global environment in an object with the same name as the dataset.
For example, if \code{dataset} is "SJTRC_minimal", it will load the data to an object
named "SJTRC_minimal" in the global environment. If the dataset is already loaded, it will not be re-loaded.
}
\details{
The \code{SJTRC_longitudinal} dataset contains 6 samples of TIRTL-seq data from the
St. Jude Tracking Study of Immune Responses Associated with COVID-19 (SJTRC).
The data is from one donor at three timepoints, isolated for either CD4+ or CD8+ T-cells.

The \code{SJTRC_minimal} dataset is a subset of the above dataset, containing only two samples:
CD8+ isolated T-cells for timepoints 1 and 2.
}
\examples{
load_example_data("SJTRC_minimal")
print(SJTRC_minimal)
summary(SJTRC_minimal)
load_example_data("SJTRC_longitudinal")
print(SJTRC_longitudinal)
summary(SJTRC_longitudinal)

}
\seealso{
Other data_loading:
\code{\link[=load_tirtlseq]{load_tirtlseq()}}
}
\concept{data_loading}
