% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tirtlseq.R
\name{load_tirtlseq}
\alias{load_tirtlseq}
\title{Load data from TIRTLseq experiments}
\usage{
load_tirtlseq(
  directory,
  chain = c("all", "paired", "alpha", "beta"),
  sep = "_",
  meta_columns = NULL,
  samples = NULL,
  clean = FALSE,
  remove_nonfunctional = FALSE,
  process = TRUE,
  pseudobulk_columns = "auto",
  paired_columns = "auto",
  n_threads = data.table::getDTthreads(),
  verbose = TRUE,
  stringsAsFactors = FALSE,
  n_max = Inf
)
}
\arguments{
\item{directory}{the directory to look in for ".tsv" or ".tsv.gz" files of TIRTLseq data}

\item{chain}{which TCR chain data to load data for -- "all" chains (alpha, beta, and paired) by default.}

\item{sep}{(optional) separator in the filename for metadata information ("_" by default)}

\item{meta_columns}{(optional) a vector of names identifying the metadata contained in the filenames,
for example \code{c("marker", "timepoint", "donor")} for files named something like "cd8_timepoint2_donor1 ... .tsv".}

\item{samples}{(optional) specific sample ids (the part of the filename before "_pseudobulk" or "_TIRTLoutput") to load. Default is NULL (loads all samples in the directory).}

\item{clean}{(optional) a TRUE/FALSE value, whether or not to "clean" the paired data by removing
excess pairs for individual alpha and beta chains (default is FALSE).}

\item{remove_nonfunctional}{whether to remove non-functional TCR chains (default is FALSE)}

\item{process}{whether to process the raw data, adding columns for alpha and beta readcounts to the
paired data frame, adding columns to the pseudobulk data to identify which chains are paired, etc.}

\item{pseudobulk_columns}{(optional) the columns of the pseudobulk .tsv(.gz) to read. Either
a list of columns or one of "auto", "all", or "minimal". "auto" (default) loads all columns except for
some redundant ones. "all" loads all columns. "minimal" loads a small number of the most important columns.}

\item{paired_columns}{(optional) the columns of the paired .tsv(.gz) to read. Either
a list of columns or one of "auto", "all", or "minimal". "auto" (default) loads all columns except for
some redundant ones. "all" loads all columns. "minimal" loads a small number of the most important columns.}

\item{n_threads}{(optional) number of CPU threads to use for reading .tsv(.gz) files}

\item{verbose}{(optional) whether to print the name of each file loaded (default is TRUE).}

\item{stringsAsFactors}{(optional) read character strings in as factors}

\item{n_max}{(optional) the maximum number of files to read in -- used mostly for testing purposes (default is Inf, i.e. read all files).}
}
\value{
The function returns a list with the following structure, containing the data from
each sample and the sample metadata.

\preformatted{
your_data_object (list)
├───meta (metadata dataframe)
└───data (list)
    └───sample_1 (list)
        ├───alpha (alpha pseudobulk dataframe)
        ├───beta (beta pseudobulk dataframe)
        └───paired (paired pseudobulk dataframe)
    ...
    └───sample_n (list)
        ├───alpha (alpha pseudobulk dataframe)
        ├───beta (beta pseudobulk dataframe)
        └───paired (paired pseudobulk dataframe)
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{load_tirtlseq()} loads paired-TCR and pseudo-bulk data from TIRTLseq experiments
from a given directory. It can also automatically assemble metadata from filenames.
}
\details{
The function expects ".tsv" (or ".tsv.gz") files. It looks for files ending in
"_pseudobulk_TRA.tsv" (alpha-chain pseudo-bulk), "_pseudobulk_TRB.tsv" (beta-chain pseudo-bulk),
and "_TIRTLoutput.tsv" (paired alpha and beta chains).

By default, the function will construct a metadata table with a row for each sample, based
on unique strings at the beginning of filenames (before "_TIRTLoutput.tsv" or similar).
If the filename contains sample metadata, then it can add multiple columns to the metadata
table with this information. For example, if a typical file looks like "cd8_timepoint2_donor1_TIRTLoutput.tsv"
and the user supplies \code{c("cell_type", "timepoint", "donor")} for \code{meta_columns} and \code{"_"} for \code{sep},
then the metadata table will look like something like this:
\preformatted{
   sample_id             cell_type   timepoint       donor     label
     <chr>               <chr>         <chr>         <chr>     <chr>
1 cd8_timepoint2_donor1    cd8       timepoint2      donor1    cell_type: cd8 | timepoint: timepoint2 | donor: donor1
2 ...
3 cd4_timepoint1_donor3    cd4       timepoint1      donor3    cell_type: cd4 | timepoint: timepoint1 | donor: donor3
}
}
\examples{
load_example_data(dataset = "SJTRC_minimal")
print(SJTRC_minimal)
summary(SJTRC_minimal)
SJTRC_minimal$meta ## metadata table
names(SJTRC_minimal$data) ## samples
SJTRC_minimal$data$cd8_tp1_v2$alpha \%>\% dplyr::as_tibble() ## alpha chain pseudobulk
SJTRC_minimal$data$cd8_tp1_v2$beta \%>\% dplyr::as_tibble() ## beta chain pseudobulk
SJTRC_minimal$data$cd8_tp1_v2$paired \%>\% dplyr::as_tibble() ## all pairs (rows duplicated if called by both pairing algorithms)
SJTRC_minimal$data$cd8_tp1_v2$paired_alt \%>\% dplyr::as_tibble() ## all pairs (each row unique)
SJTRC_minimal$info ## information

}
\seealso{
Other data_import:
\code{\link[=get_samples]{get_samples()}},
\code{\link[=load_example_data]{load_example_data()}},
\code{\link[=read_external_bulk]{read_external_bulk()}},
\code{\link[=read_external_paired]{read_external_paired()}}
}
\concept{data_import}
