% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clone_size_across_samples.R
\name{plot_clone_size_across_samples}
\alias{plot_clone_size_across_samples}
\title{Line plot of clone read fraction across multiple samples}
\usage{
plot_clone_size_across_samples(
  data,
  clones,
  chain = c("beta", "alpha"),
  pseudo = 1e-06,
  group_vec = NULL,
  sum_readFraction = TRUE,
  samples = NULL,
  return_data = FALSE,
  label_zero = FALSE,
  show_legend = TRUE,
  log_scale = TRUE,
  x_var = NULL
)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{clones}{a list of nucleotide sequences of TCR clones}

\item{chain}{the TCR chain used (default is "beta")}

\item{pseudo}{the value of the pseudocount added to all read fractions (default is 10^-6)}

\item{group_vec}{(optional) a vector of "groups" for the clones}

\item{sum_readFraction}{if TRUE, plot the sum of read fractions of clones in each group.
If FALSE, plot each clone with a separate line.}

\item{samples}{(optional) which samples to use in the plot (default is all samples)}

\item{return_data}{whether to return the data used for plotting (default is FALSE)}

\item{label_zero}{whether to label zero on the y-axis (default is FALSE)}

\item{show_legend}{whether to show the legend (default is TRUE)}

\item{log_scale}{(optional) if TRUE, use log-scale for the y-axis (default is FALSE)}

\item{x_var}{a column of metadata for grouping samples in the plot. The default is NULL,
which considers each sample its own group.}
}
\value{
a ggplot object with a line plot of clone read fractions across samples.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function creates a line plot of clone read fraction for the specified clones
across multiple samples. The function can color the lines by groups of clones.
If \code{sum_readFraction} is TRUE, the read fraction for each group will be summed
and presented in one line.
}
\examples{
folder = system.file("extdata/SJTRC_TIRTL_seq_longitudinal",
  package = "TIRTLtools")
sjtrc = load_tirtlseq(folder,
  meta_columns = c("marker", "timepoint", "version"), sep = "_",
  verbose = FALSE)

top_clones1 = sjtrc$data$cd8_tp1_v2$beta \%>\%
  dplyr::arrange(desc(readFraction)) \%>\%
  head(5) \%>\%
  magrittr::extract2("targetSequences") \%>\% as.character()
top_clones2 = sjtrc$data$cd8_tp2_v2$beta \%>\%
  dplyr::arrange(desc(readFraction)) \%>\%
  head(5) \%>\%
  magrittr::extract2("targetSequences") \%>\% as.character()

plot_clone_size_across_samples(sjtrc,
  clones = c(top_clones1, top_clones2), chain = "beta")


}
\seealso{
Other longitudinal: 
\code{\link{plot_sample_vs_sample}()}
}
\concept{longitudinal}
