% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clone_size_across_samples.R
\name{plot_clone_size_across_samples}
\alias{plot_clone_size_across_samples}
\title{Line plot of clone read fraction across multiple samples}
\usage{
plot_clone_size_across_samples(
  data,
  clones,
  chain = c("beta", "alpha"),
  pseudo = 1e-06,
  group_vec = NULL,
  sum_readFraction = TRUE,
  samples = NULL,
  return_data = FALSE,
  label_zero = FALSE,
  show_legend = TRUE,
  log_scale = TRUE,
  x_var = NULL
)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{clones}{a list of nucleotide sequences of TCR clones}

\item{chain}{the TCR chain used (default is "beta")}

\item{pseudo}{the value of the pseudocount added to all read fractions (default is 10^-6)}

\item{group_vec}{(optional) a vector of "groups" for the clones}

\item{sum_readFraction}{if TRUE, plot the sum of read fractions of clones in each group.
If FALSE, plot each clone with a separate line.}

\item{samples}{(optional) which samples to use in the plot (default is all samples)}

\item{return_data}{whether to return the data used for plotting (default is FALSE)}

\item{label_zero}{whether to label zero on the y-axis (default is FALSE)}

\item{show_legend}{whether to show the legend (default is TRUE)}

\item{log_scale}{(optional) if TRUE, use log-scale for the y-axis (default is FALSE)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other plotting: 
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
