% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clonotype_indices.R
\name{plot_clonotype_indices}
\alias{plot_clonotype_indices}
\title{Plotting of read fractions for most frequent clonotypes}
\usage{
plot_clonotype_indices(
  data,
  chain = c("beta", "alpha"),
  type_column = "auto",
  proportion_column = "auto",
  cutoffs = 10^(1:5),
  group_col = NULL,
  label_col = "Sample",
  flip = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{data}{a dataset created by \code{\link{load_tirtlseq()}} and possibly \code{\link{filter_dataset()}}}

\item{chain}{< fill in later >}

\item{type_column}{< fill in later >}

\item{proportion_column}{< fill in later >}

\item{cutoffs}{the indices used for the end of each group in the bar chart.
The default is 10^(1:5), i.e. the 10th most-frequent clone, the 100th most-frequent clone,
the 1,000th, the 10,000th, and the 100,000th.}

\item{group_col}{(optional) if supplied, a column of the metadata that will be used
to group samples}

\item{label_col}{(optional) labels for the samples}

\item{flip}{(optional) if TRUE, flip the x and y-axes (default is FALSE)}

\item{return_data}{(optional) if TRUE, return the data used to make the plot (default is FALSE)}
}
\value{
A list with two objects:

\code{$plot} - a stacked bar chart of relative frequencies of most-frequent clonotypes

\code{$data} - if return_data is TRUE, the data frame used to create the plot
}
\description{
\code{plot_clonotype_indices()} creates a stacked bar chart containing the fraction of
reads for the top 10 most frequent clonotypes, the 11th to 100th most frequent
clonotypes, the 101st to 1000th most frequent, and so on, for each sample in the dataset.
}
\details{
Fill in later.
}
\examples{
# example code


}
\seealso{
\code{\link{diversity()}}, \code{\link{plot_diversity()}}
}
