% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clusters of similar TCRs}
\usage{
plot_clusters(
  obj,
  n_clusters = 10,
  seed = 1234,
  annotation_cols = c("cluster", "source"),
  color_col = "cluster"
)
}
\arguments{
\item{obj}{an object returned by the \code{\link{cluster_tcrs}()} function}

\item{n_clusters}{the number of clusters to plot (default 10)}

\item{seed}{a number to use as the seed. Using the same number across multiple runs
allows for reproducible results.}

\item{annotation_cols}{columns to use for heatmap annotations (default is the cluster
and the source, i.e. your data or vdj-db)}

\item{color_col}{a column to use to color nodes in the network visualization (default is "cluster")}
}
\value{
Currently returns a list with three items:
\code{$umap} - a umap visualization of the TCRs from the top clusters

\code{$graph} - a graph/network visualization of TCRs from the top clusters

\code{$heatmap} - a heatmap visualization of TCRs from the top clusters
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{plot_clusters()} plots TCRs that were previously clustered by the \code{\link{cluster_tcrs}()} function.
}
\details{
The function currently returns a list with a UMAP plot, a graph/network plot, and
a heatmap.
}
\examples{
# example code
# paired = load_tirtlseq("your_directory/")
# obj = cluster_tcrs(paired)
# plot_clusters(obj)
}
\seealso{
\code{\link{cluster_tcrs}()}, \code{\link{TCRdist}()}

Other tcr_similarity:
\code{\link[=TCRdist]{TCRdist()}},
\code{\link[=TCRdist_cpp]{TCRdist_cpp()}},
\code{\link[=TCRdist_old]{TCRdist_old()}},
\code{\link[=TCRdist_to_igraph]{TCRdist_to_igraph()}},
\code{\link[=TCRdist_to_sparse_matrix]{TCRdist_to_sparse_matrix()}},
\code{\link[=add_to_tcr_network]{add_to_tcr_network()}},
\code{\link[=cluster_tcrs]{cluster_tcrs()}}
}
\concept{tcr_similarity}
