% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot the top N largest clusters of a set of TCRs}
\usage{
plot_clusters(
  obj,
  n_clusters = 10,
  seed = 1234,
  annotation_cols = c("cluster", "source"),
  color_col = "cluster"
)
}
\arguments{
\item{obj}{an object returned by the `cluster_tcrs()` function}

\item{n_clusters}{the number of clusters to plot (default 10)}

\item{seed}{a number to use as the seed. Using the same number across multiple runs
allows for reproducible results.}

\item{annotation_cols}{columns to use for heatmap annotations (default is the cluster
and the source, i.e. your data or vdj-db)}

\item{color_col}{a column to use to color nodes in the network visualization (default is "cluster")}
}
\value{
Currently returns a list with three items:
$umap - a umap visualization of the TCRs from the top clusters

$graph - a graph/network visualization of TCRs from the top clusters

$heatmap - a heatmap visualization of TCRs from the top clusters
}
\description{
`plot_clusters()` plots TCRs that were previously clustered by the `cluster_tcrs()` function.
}
\details{
The function currently returns a list with a UMAP plot, a graph/network plot, and
a heatmap.
}
\examples{
# example code
# paired = load_tirtlseq("your_directory/")
# obj = cluster_tcrs(paired)
# plot_clusters(obj)
}
\seealso{
[cluster_tcrs()], [TCRdist()]
}
