% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diversity.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Plotting of clonal diversity metrics}
\usage{
plot_diversity(
  div,
  metric = .get_all_div_metrics(),
  q = 2,
  percent = 90,
  group_col = NULL,
  label_col = "Sample",
  flip = FALSE,
  facet = FALSE,
  log_scale = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{div}{a list created by the `diversity` function with diversity metrics for each sample}

\item{metric}{the diversity metric to use (e.g. shannon, simpson, etc.)}

\item{q}{(optional) for 'renyi' and 'hill' metrics, the order q of the diversity index}

\item{percent}{(optional) for 'dXX' metric, the percentage 'XX' between 0 and 100}
}
\description{
The `plot_diversity()` function returns...
}
\details{
This function...
}
