% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diversity.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Bar plot of clonal diversity metrics}
\usage{
plot_diversity(
  div,
  metric = get_all_div_metrics(),
  q = 2,
  percent = 90,
  group_col = NULL,
  label_col = "Sample",
  flip = FALSE,
  facet = FALSE,
  log_scale = FALSE,
  samples = NULL,
  return_data = FALSE,
  color_scheme = NULL,
  x_var = NULL
)
}
\arguments{
\item{div}{a list created by the \code{\link{diversity}()} function with diversity metrics for each sample}

\item{metric}{the diversity metric to use (e.g. shannon, simpson, etc.)}

\item{q}{(optional) for 'renyi' and 'hill' metrics, the order q of the diversity index}

\item{percent}{(optional) for 'dXX' metric, the percentage 'XX' between 0 and 100}

\item{group_col}{(optional) if supplied, a column of the metadata that will be used
to group samples}

\item{label_col}{(optional) labels for the samples}

\item{flip}{(optional) if TRUE, flip the x and y-axes (default is FALSE)}

\item{facet}{(optional) if TRUE, plot with separate facets, or sub-plots for each group (default is FALSE)}

\item{log_scale}{(optional) if TRUE, use log-scale for the y-axis (default is FALSE)}

\item{samples}{(optional) the samples to include in the plot (default is all samples)}

\item{return_data}{(optional) if TRUE, return the data used to make the plot (default is FALSE)}

\item{color_scheme}{(optional) a color scheme for the plot}

\item{x_var}{(optional) the variable to show on the x-axis, if other than "sample_id"}
}
\value{
A list with two objects:

\code{$plot} - a ggplot object with the plot of the requested diversity metric

\code{$data} - if return_data is TRUE, the data frame used to make the plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The \code{plot_diversity()} plots the requested clonal diversity metric
}
\details{
This function can plot a variety of clonal diversity metrics for a dataset (richness,
Simpson diversity index, Shannon-Wiener index, etc.). See \code{\link{get_all_div_metrics}()} for
all available options. By default it return a barplot with one bar for each sample in
the dataset. If a grouping column (of the metadata) is supplied, then samples will be grouped and
bar heights will reflect the average diversity metric across the group, with error bars
showing the standard deviation.
}
\examples{
# example code
# data = load_tirtlseq("your_directory/")
# div = diversity(data)
# plot_diversity(div, metric = "richness")

}
\seealso{
\code{\link{diversity}()}, \code{\link{get_all_div_metrics}()}

Other plotting: 
\code{\link{plot_clone_size_across_samples}()},
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}

Other diversity: 
\code{\link{calculate_diversity}()}
}
\concept{diversity}
\concept{plotting}
