% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_usage.R
\name{plot_gene_usage}
\alias{plot_gene_usage}
\title{Plotting of V- or J- gene usage}
\usage{
plot_gene_usage(
  data,
  chain = c("paired", "alpha", "beta"),
  group_col = NULL,
  gene = c("va", "vb", "ja", "jb", "v", "j"),
  n_max = 25,
  value_type = c("auto", "readFraction", "readCount", "n_wells", "readCount_max",
    "readCount_median", "avg", "n"),
  agg_func = c("sum", "max"),
  log_scale = FALSE,
  style = c("barplot", "boxplot")
)
}
\arguments{
\item{data}{either a single data frame containing paired TCRs from TIRTL-seq output or a list of
data frames for many experiments}

\item{gene}{the gene of the TCRs to plots}

\item{n_max}{the maximum number of V-segments to plot. The top \code{n_max} most common V-segments will be plotted}

\item{groups}{(optional) when data is a list of data frames, this may be a vector of the same
length as data with labels to group the experiments.}
}
\value{
a ggplot object with a bar chart of the most common V- or J- segments.
}
\description{
\code{plot_gene_usage()} creates a bar chart of gene segment usage for the most common
V- or J- genes.
}
\details{
Fill in later, after updating/finalizing function.
}
\examples{
# example here
}
