% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_reads.R
\name{plot_n_reads}
\alias{plot_n_reads}
\title{Bar plot of the number of single-chain reads for each sample}
\usage{
plot_n_reads(
  data,
  chain = c("both", "beta", "alpha"),
  samples = NULL,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{chain}{the TCR chain to plot (default is both "alpha" and "beta")}

\item{samples}{(optional) which samples to plot (default is all samples)}

\item{color_scheme}{a color scheme for the plot}
}
\description{
Bar plot of the number of single-chain reads for each sample
}
\seealso{
Other plotting: 
\code{\link{plot_clone_size_across_samples}()},
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
