% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_reads.R
\name{plot_n_reads}
\alias{plot_n_reads}
\title{Bar plot of the number of single-chain reads for each sample}
\usage{
plot_n_reads(
  data,
  chain = c("both", "beta", "alpha"),
  samples = NULL,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{chain}{the TCR chain to plot (default is both "alpha" and "beta")}

\item{samples}{(optional) which samples to plot (default is all samples)}

\item{color_scheme}{a color scheme for the plot}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function plots the total number of alpha and/or beta reads (default is both)
for each sample.
}
\examples{
folder = system.file("extdata/SJTRC_TIRTL_seq_longitudinal", package = "TIRTLtools")
ts_data = load_tirtlseq(folder, meta_columns = c("marker", "timepoint", "version"), sep = "_", verbose = FALSE)

plot_n_reads(ts_data)


}
\seealso{
Other qc: 
\code{\link{get_all_tcrs}()},
\code{\link{get_pair_stats}()},
\code{\link{get_paired_by_read_fraction_range}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{summarize_data}()}
}
\concept{qc}
