% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_num_partners.R
\name{plot_num_partners}
\alias{plot_num_partners}
\title{Stacked bar plot of the fraction of alpha/beta chains with different numbers of partners}
\usage{
plot_num_partners(
  data,
  group_col = NULL,
  fraction = TRUE,
  include_non_functional = FALSE,
  max_partners = 5,
  samples = NULL,
  return_data = FALSE,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{group_col}{a column of the metadata to use to group multiple samples into one bar plot
The default is NULL, which uses the sample id.}

\item{fraction}{whether to plot the fraction of chains or the total number of chains
(default is TRUE, i.e. plot fractions)}

\item{include_non_functional}{whether to include chains with non-functional
cdr3 sequences when tabulating the output.}

\item{max_partners}{the maximum number of partners, N, to include in the plots.
All chains with more than N partners will be grouped together under the ">N" category.}

\item{samples}{(optional) which samples to plot}

\item{return_data}{if TRUE, return the data used to make the plots}

\item{color_scheme}{the color scheme to use for the plot}
}
\value{
A bar chart (ggplot object) with facets (sub-plots) for each sample.
If \code{return_data} is TRUE, the data frame used to create the plot is returned instead.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{plot_num_partners()} creates bar plots for alpha and beta chains showing how many
partners they are paired with by the MAD-HYPE and/or T-shell algorithms.
}
\details{
For each sample, the function creates stacked bar charts for alpha and beta chains,
showing the proportion of them (among all called pairs) that are paired with 1 chain,
2 chains, 3 chains, etc.
}
\examples{
load_example_data(dataset = "SJTRC_longitudinal")

plot_num_partners(SJTRC_longitudinal)


}
\seealso{
\code{\link{identify_non_functional_seqs}()}

Other qc:
\code{\link[=get_all_tcrs]{get_all_tcrs()}},
\code{\link[=get_pair_stats]{get_pair_stats()}},
\code{\link[=get_paired_by_read_fraction_range]{get_paired_by_read_fraction_range()}},
\code{\link[=plot_n_paired_step]{plot_n_paired_step()}},
\code{\link[=plot_n_reads]{plot_n_reads()}},
\code{\link[=plot_paired]{plot_paired()}},
\code{\link[=plot_paired_by_read_fraction_range]{plot_paired_by_read_fraction_range()}},
\code{\link[=plot_paired_vs_rank]{plot_paired_vs_rank()}},
\code{\link[=plot_pairs_with_eachother]{plot_pairs_with_eachother()}},
\code{\link[=plot_ranks]{plot_ranks()}},
\code{\link[=plot_read_fraction_vs_pair_status]{plot_read_fraction_vs_pair_status()}},
\code{\link[=plot_sample_overlap]{plot_sample_overlap()}},
\code{\link[=qc_report]{qc_report()}},
\code{\link[=summarize_data]{summarize_data()}}
}
\concept{qc}
