% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_num_partners.R
\name{plot_num_partners}
\alias{plot_num_partners}
\title{Plot the number of partners for each single chain that is matched to another chain.}
\usage{
plot_num_partners(
  data,
  group_col = NULL,
  fraction = TRUE,
  include_non_functional = FALSE,
  max_partners = 5,
  samples = NULL,
  return_data = FALSE,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a dataset created by \code{\link{load_tirtlseq}()} and possibly \code{\link{filter_dataset}()}}

\item{group_col}{a column of the metadata to use to group multiple samples into one bar plot}

\item{fraction}{whether to plot the fraction of chains or the total number of chains
(default is TRUE, i.e. plot fractions)}

\item{include_non_functional}{whether to include chains with non-functional
cdr3 sequences when tabulating the output.}

\item{max_partners}{the maximum number of partners, N, to include in the plots.
All chains with more than N partners will be grouped together under the ">N" category.}

\item{return_data}{if TRUE, return the data used to make the plots}
}
\value{
Either a bar chart (ggplot object) with facets (sub-plots) for each sample or
a list with two objects:

\code{$plot} the plot referenced above

\code{$data} the data used to create the plot
}
\description{
\code{plot_num_partners()} creates bar plots for alpha and beta chains showing how many
partners they are paired with by the MAD-HYPE and/or T-shell algorithms.
}
\details{
For each sample, the function creates stacked bar charts for alpha and beta chains,
showing the proportion of them (among all called pairs) that are paired with 1 chain,
2 chains, 3 chains, etc.
}
\examples{
# example code


}
\seealso{
\code{\link{identify_non_functional_seqs}()}
}
