% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired.R
\name{plot_paired}
\alias{plot_paired}
\title{Stacked bar plot of the number alpha/beta chains paired by each pairing algorithm}
\usage{
plot_paired(
  data,
  chain = c("paired", "alpha", "beta"),
  samples = NULL,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqData object or a data frame created using \code{\link{get_pair_stats}()}}

\item{chain}{the TCR chain to plot (default is "paired")}

\item{samples}{(optional) the samples to include in the plot (default is all)}

\item{color_scheme}{(optional) the color scheme for the plot}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other plotting: 
\code{\link{plot_clone_size_across_samples}()},
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
