% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired_by_read_fraction_range.R
\name{plot_paired_by_read_fraction_range}
\alias{plot_paired_by_read_fraction_range}
\title{Bar plot of the fraction of single chains that were paired within different
read fraction ranges for each sample.}
\usage{
plot_paired_by_read_fraction_range(
  data,
  chain = c("beta", "alpha"),
  cutoffs = 10^(-6:-1),
  freq = FALSE,
  samples = NULL,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqData object or a data frame created using \code{\link{get_paired_by_read_fraction_range}()}}

\item{chain}{the TCR chain to plot (default is "beta")}

\item{cutoffs}{a vector of cutoffs for the read fraction ranges}

\item{freq}{if TRUE, plot the number of pairs, if FALSE plot the fraction paired
(default is FALSE, plot the fraction paired)}

\item{samples}{(optional) the samples to include in the plot}

\item{color_scheme}{(optional) the color scheme to use in the plot}
}
\description{
Bar plot of the fraction of single chains that were paired within different
read fraction ranges for each sample.
}
\seealso{
Other plotting: 
\code{\link{plot_clone_size_across_samples}()},
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
