% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired_vs_rank.R
\name{plot_paired_vs_rank}
\alias{plot_paired_vs_rank}
\title{A step plot of the cumulative number of paired/unpaired alpha/beta chains among the most frequent chains}
\usage{
plot_paired_vs_rank(
  data,
  sample,
  y_axis = c("n_not_paired", "n_paired"),
  chain = c("both", "beta", "alpha"),
  n_max = 100,
  by_method = TRUE,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{sample}{the sample to plot (either by number or sample id)}

\item{y_axis}{whether to plot the cumulative number of paired or un-paired single-chains (default is "n_not_paired")}

\item{chain}{the TCR chain to plot (default is both alpha and beta)}

\item{n_max}{the number of most frequent single-chains to plot}

\item{color_scheme}{(optional) the color scheme for the plot}
}
\value{
A ggplot object with the plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The function creates a stepped line plot of the cumulative number of paired/unpaired alpha
or beta chains (default is both) among the \code{n_max} most frequent chains.
}
\examples{
load_example_data(dataset = "SJTRC_longitudinal")

plot_paired_vs_rank(SJTRC_longitudinal, sample = 1, y_axis = "n_not_paired", chain = "both", n_max = 100)

plot_paired_vs_rank(SJTRC_longitudinal, sample = 1, y_axis = "n_paired", chain = "both", n_max = 100)


}
\seealso{
Other qc:
\code{\link[=get_all_tcrs]{get_all_tcrs()}},
\code{\link[=get_pair_stats]{get_pair_stats()}},
\code{\link[=get_paired_by_read_fraction_range]{get_paired_by_read_fraction_range()}},
\code{\link[=plot_n_reads]{plot_n_reads()}},
\code{\link[=plot_num_partners]{plot_num_partners()}},
\code{\link[=plot_paired]{plot_paired()}},
\code{\link[=plot_paired_by_read_fraction_range]{plot_paired_by_read_fraction_range()}},
\code{\link[=plot_pairs_with_eachother]{plot_pairs_with_eachother()}},
\code{\link[=plot_ranks]{plot_ranks()}},
\code{\link[=plot_read_fraction_vs_pair_status]{plot_read_fraction_vs_pair_status()}},
\code{\link[=plot_sample_overlap]{plot_sample_overlap()}},
\code{\link[=summarize_data]{summarize_data()}}
}
\concept{qc}
