% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired_vs_rank.R
\name{plot_paired_vs_rank}
\alias{plot_paired_vs_rank}
\title{Line plot of the number of unpaired single-chains for the N most frequent single-chains}
\usage{
plot_paired_vs_rank(
  data,
  sample = 1,
  y_axis = c("n_not_paired", "n_paired"),
  chain = c("both", "beta", "alpha"),
  n_max = 100,
  color_scheme = NULL
)
}
\description{
Line plot of the number of unpaired single-chains for the N most frequent single-chains
}
\seealso{
Other plotting: 
\code{\link{plot_clone_size_across_samples}()},
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
