% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired_vs_rank.R
\name{plot_paired_vs_rank}
\alias{plot_paired_vs_rank}
\title{A stepped plot of the cumulative number of paired (or unpaired) single-chains for the N most frequent single-chains}
\usage{
plot_paired_vs_rank(
  data,
  sample = 1,
  y_axis = c("n_not_paired", "n_paired"),
  chain = c("both", "beta", "alpha"),
  n_max = 100,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{sample}{the sample to plot}

\item{y_axis}{whether to plot the cumulative number of paired or un-paired single-chains (default is "n_not_paired")}

\item{chain}{the TCR chain to plot (default is both alpha and beta)}

\item{n_max}{the number of most frequent single-chains to plot}

\item{color_scheme}{(optional) the color scheme for the plot}
}
\description{
A stepped plot of the cumulative number of paired (or unpaired) single-chains for the N most frequent single-chains
}
\seealso{
Other plotting: 
\code{\link{plot_clone_size_across_samples}()},
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
