% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs_with_eachother.R
\name{plot_pairs_with_eachother}
\alias{plot_pairs_with_eachother}
\title{A connected point plot of read fraction vs. rank for the most frequent alpha/beta chains}
\usage{
plot_pairs_with_eachother(
  data,
  sample,
  n_max = 100,
  show_num_partners = FALSE,
  color_scheme = NULL
)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{sample}{the sample to plot (either by number or sample id)}

\item{n_max}{the number of most frequent single-chains to plot}

\item{show_num_partners}{whether to show the number of partners for each single-chain (default is FALSE)}

\item{color_scheme}{(optional) the color scheme for the plot}
}
\value{
A ggplot object with the plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function creates a point plot of read fraction vs. rank for the N most frequent alpha (left, mirrored)
and beta (right) chains with lines between alpha and beta chains indicating a
pair and a cross indicating an unpaired single-chain.
}
\examples{
folder = system.file("extdata/SJTRC_TIRTL_seq_longitudinal", package = "TIRTLtools")
ts_data = load_tirtlseq(folder, meta_columns = c("marker", "timepoint", "version"), sep = "_", verbose = FALSE)

plot_pairs_with_eachother(ts_data, sample = 1, n_max = 100)

plot_pairs_with_eachother(ts_data, sample = 1, n_max = 100, show_num_partners = TRUE)


}
\seealso{
Other qc: 
\code{\link{get_all_tcrs}()},
\code{\link{get_pair_stats}()},
\code{\link{get_paired_by_read_fraction_range}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{summarize_data}()}
}
\concept{qc}
