% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs_with_eachother.R
\name{plot_pairs_with_eachother}
\alias{plot_pairs_with_eachother}
\title{Point plot of read fraction vs. rank for the N most frequent alpha (left, mirrored)
and beta (right) chains with lines between alpha and beta chains indicating a
pair and a cross indicating an unpaired single-chain}
\usage{
plot_pairs_with_eachother(
  data,
  sample = 1,
  n_max = 100,
  show_num_partners = FALSE,
  color_scheme = NULL
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other plotting: 
\code{\link{plot_clone_size_across_samples}()},
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
