% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranks.R, R/rank_plot.R
\name{plot_ranks}
\alias{plot_ranks}
\alias{rank_plot}
\title{Line plot of clonotype rank vs. read fraction for each sample}
\usage{
plot_ranks(
  data,
  chain = c("beta", "alpha"),
  column = "readFraction",
  color_scheme = NULL
)

rank_plot(...)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{chain}{the chain to plot, "alpha" or "beta" (default is "beta")}

\item{column}{the variable to plot (default is "readFraction")}

\item{color_scheme}{(optional) the color scheme for the plot}
}
\value{
a ggplot object with the rank plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{plot_ranks()} returns a line plot of TCR rank vs. read fraction for a given sample.
}
\examples{
folder = system.file("extdata/SJTRC_TIRTL_seq_longitudinal", package = "TIRTLtools")
ts_data = load_tirtlseq(folder, meta_columns = c("marker", "timepoint", "version"), sep = "_", verbose = FALSE)

plot_ranks(ts_data, chain = "beta")


}
\seealso{
Other qc: 
\code{\link{get_all_tcrs}()},
\code{\link{get_pair_stats}()},
\code{\link{get_paired_by_read_fraction_range}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_overlap}()},
\code{\link{summarize_data}()}
}
\concept{qc}
