% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_fraction_vs_pair_status.R
\name{plot_read_fraction_vs_pair_status}
\alias{plot_read_fraction_vs_pair_status}
\title{Alternative point plot of read fraction vs. rank for the N most frequent alpha
and beta chains, with a cross indicating an unpaired single-chain}
\usage{
plot_read_fraction_vs_pair_status(
  data,
  sample = 1,
  chain = c("both", "beta", "alpha"),
  n_max = 100,
  show_num_partners = F,
  color_scheme = NULL
)
}
\description{
Alternative point plot of read fraction vs. rank for the N most frequent alpha
and beta chains, with a cross indicating an unpaired single-chain
}
\seealso{
Other plotting: 
\code{\link{plot_clone_size_across_samples}()},
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_sample_overlap}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
