% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_fraction_vs_pair_status.R
\name{plot_read_fraction_vs_pair_status}
\alias{plot_read_fraction_vs_pair_status}
\title{Alternative point plot of read fraction vs. rank for the N most frequent alpha
and beta chains, with a cross indicating an unpaired single-chain}
\usage{
plot_read_fraction_vs_pair_status(
  data,
  sample = 1,
  chain = c("both", "beta", "alpha"),
  n_max = 100,
  show_num_partners = F,
  color_scheme = NULL
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other qc: 
\code{\link{get_all_tcrs}()},
\code{\link{get_pair_stats}()},
\code{\link{get_paired_by_read_fraction_range}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_sample_overlap}()},
\code{\link{summarize_data}()}
}
\concept{qc}
