% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_overlap.R, R/sample_overlap.R
\name{plot_sample_overlap}
\alias{plot_sample_overlap}
\alias{sample_overlap}
\title{Plot the overlap/agreement between samples (in terms of most frequent clones)}
\usage{
plot_sample_overlap(
  data,
  chain = c("paired", "alpha", "beta"),
  n_seq = 200,
  show_row_names = FALSE,
  show_column_names = FALSE,
  label_col = "Sample",
  title = "",
  return_data = FALSE,
  cluster_rows = TRUE,
  cluster_cols = TRUE
)

sample_overlap(...)
}
\arguments{
\item{data}{the dataset, an object loaded using the \code{\link{load_tirtlseq}()} function}

\item{chain}{which chain to plot: either paired or alpha-/beta-pseudobulk. (default "paired")}

\item{n_seq}{the number of most frequent TCR sequences to use (default 200)}

\item{show_row_names}{whether to show row names for the heatmap (default FALSE)}

\item{show_column_names}{whether to show column names for the heatmap (default FALSE)}

\item{label_col}{a column of the metadata to use as labels for rows and columns (default "Sample", uses the sample_id)}

\item{title}{a title for the heatmap}

\item{return_data}{whether to return the data used for plotting (a matrix with the overlap values) instead of a heatmap (default is FALSE)}

\item{cluster_rows}{whether to cluster the rows of the heatmap (default TRUE)}

\item{cluster_cols}{whether to cluster the columns of the heatmap (default TRUE)}
}
\value{
A heatmap with hierarchically clustered rows and columns showing the number of
TCRs shared between each pair of samples, among their top N most frequent TCRs.

If \code{return_data} is TRUE, a matrix of overlap values will be returned instead.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function returns a heatmap showing the overlap among the most frequent
\code{n_seq} TCRs (default is n_seq=200) among pairs of samples in a dataset.
}
\examples{
folder = system.file("extdata/SJTRC_TIRTL_seq_longitudinal", package = "TIRTLtools")
ts_data = load_tirtlseq(folder, meta_columns = c("marker", "timepoint", "version"), sep = "_", verbose = FALSE)

plot_sample_overlap(ts_data, chain = "beta")
plot_sample_overlap(ts_data, chain = "alpha")
plot_sample_overlap(ts_data, chain = "paired")

}
\seealso{
\code{\link{load_tirtlseq}()}

Other qc: 
\code{\link{get_all_tcrs}()},
\code{\link{get_pair_stats}()},
\code{\link{get_paired_by_read_fraction_range}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{summarize_data}()}
}
\concept{qc}
