% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_overlap.R, R/sample_overlap.R
\name{plot_sample_overlap}
\alias{plot_sample_overlap}
\alias{diversity}
\title{Plot the overlap/agreement between samples (in terms of most frequent clones)}
\usage{
plot_sample_overlap(
  data,
  chain = c("paired", "alpha", "beta"),
  n_seq = 200,
  show_row_names = FALSE,
  show_column_names = FALSE,
  label_col = "Sample",
  title = "",
  return_data = FALSE,
  cluster_rows = TRUE,
  cluster_cols = TRUE
)

sample_overlap(...)
}
\arguments{
\item{data}{the dataset, an object loaded using the \code{\link{load_tirtlseq}()} function}

\item{chain}{which chain to plot: either paired or alpha-/beta-pseudobulk. (default "paired")}

\item{n_seq}{the number of most frequent TCR sequences to use (default 200)}

\item{show_row_names}{whether to show row names for the heatmap (default FALSE)}

\item{show_column_names}{whether to show column names for the heatmap (default FALSE)}

\item{label_col}{a column of the metadata to use as labels for rows and columns (default "Sample", uses the sample_id)}

\item{title}{a title for the heatmap}
}
\value{
A heatmap with hierarchically clustered rows and columns showing the number of
TCRs shared between each pair of samples, among their top N most frequent TCRs.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{plot_sample_overlap()} returns a heatmap showing the overlap in most frequent TCRs
among pairs of samples in a dataset
}
\details{
The function takes the top N most frequent TCRs found in each dataset (default 200) and
compares their overlap between samples.
}
\examples{
# example code
# data = load_tirtlseq("your_directory/")
# plot_sample_overlap(data, chain = "paired", n_seq=200)

}
\seealso{
\code{\link{load_tirtlseq}()}

Other plotting: 
\code{\link{plot_clone_size_across_samples}()},
\code{\link{plot_clonotype_indices}()},
\code{\link{plot_clusters}()},
\code{\link{plot_diversity}()},
\code{\link{plot_n_reads}()},
\code{\link{plot_num_partners}()},
\code{\link{plot_paired}()},
\code{\link{plot_paired_by_read_fraction_range}()},
\code{\link{plot_paired_vs_rank}()},
\code{\link{plot_pairs_with_eachother}()},
\code{\link{plot_ranks}()},
\code{\link{plot_read_fraction_vs_pair_status}()},
\code{\link{plot_sample_vs_sample}()}
}
\concept{plotting}
