% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_vs_sample.R
\name{plot_sample_vs_sample}
\alias{plot_sample_vs_sample}
\title{Scatterplot of TCR frequency between two samples}
\usage{
plot_sample_vs_sample(
  data1,
  data2,
  type_column = "auto",
  value_type = c("auto", "readFraction", "readCount", "n_wells", "readCount_max",
    "readCount_median", "avg", "n"),
  log2_cutoff = 1.5
)
}
\arguments{
\item{data1}{a dataframe for the first sample}

\item{data2}{a dataframe for the second sample}

\item{log2_cutoff}{the log2 fold-change cutoff to call a TCR up- or down-regulated (default 1.5)}
}
\value{
A scatterplot (ggplot object) with read frequencies (proportions), colored by whether each
TCR is up-regulated, down-regulated, or neither, given the log2 fold-change cutoff.
}
\description{
`plot_sample_vs_sample()` returns a scatterplot of read frequencies of TCRs between two samples
}
\details{
The function labels each TCR as up-regulated, down-regulated, or stable, based on the
log2 fold-change cutoff supplied (default 1.5).
}
\examples{
# example code


}
\seealso{
[func1()], [func2()], and [func3()] for similar functions
}
