% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_tcrdist.R
\name{prep_for_tcrdist}
\alias{prep_for_tcrdist}
\title{Prepare paired TCRs for TCRdist calculation}
\usage{
prep_for_tcrdist(
  df,
  params = NULL,
  remove_duplicate_tcrs = FALSE,
  remove_MAIT = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{df}{a data frame with one TCR per row. It must have the columns "va", "vb", "cdr3a", and "cdr3b"}

\item{params}{(optional) a data frame with permissible amino acids and V-segments.
By default, the \code{\link{params}} data frame is used.}

\item{remove_duplicate_tcrs}{(optional) whether or not to remove duplicate TCRs (default is FALSE, do not remove duplicates).}

\item{remove_MAIT}{(optional) whether or not to remove MAIT cells from the data frame (default is TRUE).}

\item{verbose}{(optional) whether to print progress of the function (default is TRUE).}
}
\value{
Returns a paired TCR data frame that is compatible with TCRdist python scripts.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function prepares a paired TCR data frame for running TCRdist,
adding alleles to V-segments (i.e. "*01") if they are missing and removing
TCRs with unrecognized V-segments or non-functional CDR3 sequences.
}
\seealso{
Other data_processing:
\code{\link[=clean_pairs]{clean_pairs()}},
\code{\link[=combine_bulk_and_paired_data]{combine_bulk_and_paired_data()}},
\code{\link[=filter_duplicate_tcrs]{filter_duplicate_tcrs()}},
\code{\link[=filter_mait]{filter_mait()}},
\code{\link[=filter_nonfunctional_TCRs]{filter_nonfunctional_TCRs()}},
\code{\link[=filter_short_cdr3s]{filter_short_cdr3s()}},
\code{\link[=filter_v_alleles]{filter_v_alleles()}},
\code{\link[=make_tcr_schema]{make_tcr_schema()}},
\code{\link[=read_external_bulk]{read_external_bulk()}},
\code{\link[=read_external_paired]{read_external_paired()}},
\code{\link[=remove_duplicates]{remove_duplicates()}}
}
\concept{data_processing}
