% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_tcrdist.R
\name{prep_for_tcrdist}
\alias{prep_for_tcrdist}
\title{Prepares paired TCRs for TCRdist calculation, adding alleles to V-segments
(i.e. "*01") if they are missing and removing TCRs with unrecognized V-segments
or non-functional CDR3 sequences}
\usage{
prep_for_tcrdist(df, params = NULL)
}
\description{
Prepares paired TCRs for TCRdist calculation, adding alleles to V-segments
(i.e. "*01") if they are missing and removing TCRs with unrecognized V-segments
or non-functional CDR3 sequences
}
\seealso{
Other data_processing: 
\code{\link{TIRTL_process}()},
\code{\link{add_single_chain_data}()},
\code{\link{clean_pairs}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{remove_duplicates}()}
}
\concept{data_processing}
