% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_TIRTLseq.R
\name{process_TIRTLseq}
\alias{process_TIRTLseq}
\title{Run data processing functions on a TIRTLseqDataSet object}
\usage{
process_TIRTLseq(data, clean = FALSE, remove_nonfunctional = FALSE)
}
\arguments{
\item{data}{a TIRTLseqDataSet object}

\item{clean}{(optional) a TRUE/FALSE value, whether or not to "clean" the paired data by removing
excess pairs for individual alpha and beta chains (default is FALSE).}

\item{remove_nonfunctional}{whether to remove non-functional TCR chains (default is FALSE)}
}
\value{
a TIRTLseqDataSet object with annotated and (optionally) cleaned data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function has been deprecated because its functionality has been added to the load_tirtlseq() function

This function runs annotation and data cleaning functions on a
TIRTLseqDataSet object. Specifically, it calls the functions \code{add_single_chain_data()},
\code{identify_paired()}, \code{identify_non_functional_seqs()}, and (optionally) \code{clean_pairs()}.
}
\seealso{
Other deprecated:
\code{\link[=add_single_chain_data]{add_single_chain_data()}},
\code{\link[=identify_non_functional_seqs]{identify_non_functional_seqs()}},
\code{\link[=identify_paired]{identify_paired()}}
}
\concept{deprecated}
