% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_external_bulk.R
\name{read_external_bulk}
\alias{read_external_bulk}
\title{Read and process bulk single-chain TCR-seq data}
\usage{
read_external_bulk(path, format = "MiXCR")
}
\arguments{
\item{path}{the path to the data file}

\item{format}{the data format. Currently only "MiXCR" is supported.}
}
\value{
A list containing the following slots:
\itemize{
\item df - a data frame with a few columns modified and renamed
\item df_raw - the original data, un-modified
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function reads and processes bulk single-chain TCR-sequencing data from a non-TIRTLseq assay.
Currently, only MiXCR format is supported.
}
\details{
Supported data types:
\itemize{
\item \code{"MiXCR"} - "<sample_name>_TRA.tsv" and "<sample_name>_TRB.tsv"
}
}
\seealso{
Other data_import:
\code{\link[=load_example_data]{load_example_data()}},
\code{\link[=load_tirtlseq]{load_tirtlseq()}},
\code{\link[=read_external_paired]{read_external_paired()}}
}
\concept{data_import}
