% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_external_bulk.R
\name{read_external_bulk}
\alias{read_external_bulk}
\title{Read and process bulk single-chain TCR-seq data}
\usage{
read_external_bulk(path, format = "MiXCR")
}
\arguments{
\item{path}{the path to the data file}

\item{format}{the data format. Currently only "MiXCR" is supported.}
}
\value{
A list containing the following slots:
\itemize{
\item df - a data frame with a few columns modified and renamed
\item df_raw - the original data, un-modified
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function reads and processes bulk single-chain TCR-sequencing data from a non-TIRTLseq assay.
Currently, only MiXCR format is supported.
}
\details{
Supported data types:
\itemize{
\item \code{"MiXCR"} - "<sample_name>_TRA.tsv" and "<sample_name>_TRB.tsv"
}
}
\seealso{
Other data_processing:
\code{\link[=TIRTL_process]{TIRTL_process()}},
\code{\link[=add_single_chain_data]{add_single_chain_data()}},
\code{\link[=clean_pairs]{clean_pairs()}},
\code{\link[=combine_bulk_and_paired_data]{combine_bulk_and_paired_data()}},
\code{\link[=filter_duplicate_tcrs]{filter_duplicate_tcrs()}},
\code{\link[=filter_mait]{filter_mait()}},
\code{\link[=filter_nonfunctional_TCRs]{filter_nonfunctional_TCRs()}},
\code{\link[=filter_short_cdr3s]{filter_short_cdr3s()}},
\code{\link[=filter_v_alleles]{filter_v_alleles()}},
\code{\link[=identify_non_functional_seqs]{identify_non_functional_seqs()}},
\code{\link[=identify_paired]{identify_paired()}},
\code{\link[=make_tcr_schema]{make_tcr_schema()}},
\code{\link[=prep_for_tcrdist]{prep_for_tcrdist()}},
\code{\link[=read_external_paired]{read_external_paired()}},
\code{\link[=remove_duplicates]{remove_duplicates()}}
}
\concept{data_processing}
