% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_external_paired.R
\name{read_external_paired}
\alias{read_external_paired}
\title{Read and process single-cell paired-chain TCR-seq data}
\usage{
read_external_paired(
  path,
  format = c("auto", "10X", "ParseBio"),
  id_cols = make_tcr_schema(features = c("v", "j", "cdr3_aa", "cdr3_nt"), second_alpha =
    FALSE),
  multi = FALSE,
  separate_rows = TRUE,
  productive_only = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{path}{the path to the data file}

\item{format}{the format of the data, either \code{"10X"} or \code{"ParseBio"} or \code{"auto"}. If \code{"auto"} (default), the function will try to decipher which technology the data file was created by.}

\item{id_cols}{a vector of column names to be used to define a clone (e.g. \code{c("va","vb","ja","jb","alpha_nuc", "beta_nuc", "cdr3a", "cdr3b")}). This can be produced automatically using the \code{\link{make_tcr_schema}()} function.}

\item{multi}{If \code{FALSE} (default), select only the best two alpha chains for each beta chain
when processing the data and creating a data frame with paired TCRs. If \code{TRUE}, keep all alphas.}

\item{separate_rows}{If \code{TRUE}, when there are multiple alpha chains
paired with one beta chain, put each pair in a separate row in the output data frame. If
\code{FALSE}, add second alpha chain in extra columns on the same row.}

\item{productive_only}{If \code{TRUE}, keep only "productive" chains}

\item{verbose}{If \code{TRUE}, print messages.}
}
\value{
A list containing the following slots:
\itemize{
\item df_pairs_complete - (data frame) paired receptors - one row for each receptor
(excluding those missing an alpha or beta chain)
\item df_pairs - (data frame) with paired receptors - one row for each receptor
(including those missing an alpha or beta chain)
\item df_pairs_long - (data frame) with paired receptors - one row for each cell
(including those missing an alpha or beta chain)
\item df_pairs_long_complete - (data frame) with paired receptors - one row for each cell
(excluding those missing an alpha or beta chain)
\item df_raw - (data frame) un-edited input data
\item chain_df - (data frame) summary of total number of each chain in input data
\item barcode_df - (data frame) summary of number of chains found in each cell
\item id_cols - (character vector) columns used as IDs to uniquely define receptor pairs
\item n_cells_total -(integer) total number of cells
\item n_cells_complete - (integer) number of cells with both chains
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function reads and processes paired TCR-sequencing data from a non-TIRTLseq assay.
Currently 10X and Parse Biosciences data are supported.
}
\details{
Supported data types:
\itemize{
\item \code{"10X"} - "filtered_contig_annotations.csv" or "all_contig_annotations.csv" outputs from Cell Ranger
(https://www.10xgenomics.com/support/software/cell-ranger/7.2/analysis/outputs/cr-5p-outputs-overview-vdj)
\item \code{"ParseBio"} - "tcr_annotation_airr.tsv" output from Trailmaker
}
}
\seealso{
Other data_processing:
\code{\link[=TIRTL_process]{TIRTL_process()}},
\code{\link[=add_single_chain_data]{add_single_chain_data()}},
\code{\link[=clean_pairs]{clean_pairs()}},
\code{\link[=combine_bulk_and_paired_data]{combine_bulk_and_paired_data()}},
\code{\link[=filter_duplicate_tcrs]{filter_duplicate_tcrs()}},
\code{\link[=filter_mait]{filter_mait()}},
\code{\link[=filter_nonfunctional_TCRs]{filter_nonfunctional_TCRs()}},
\code{\link[=filter_short_cdr3s]{filter_short_cdr3s()}},
\code{\link[=filter_v_alleles]{filter_v_alleles()}},
\code{\link[=identify_non_functional_seqs]{identify_non_functional_seqs()}},
\code{\link[=identify_paired]{identify_paired()}},
\code{\link[=make_tcr_schema]{make_tcr_schema()}},
\code{\link[=prep_for_tcrdist]{prep_for_tcrdist()}},
\code{\link[=read_external_bulk]{read_external_bulk()}},
\code{\link[=remove_duplicates]{remove_duplicates()}}
}
\concept{data_processing}
