% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Removes duplicate paired TCRs}
\usage{
remove_duplicates(data)
}
\arguments{
\item{data}{either a single data frame (paired chain) or a list
of data frames}
}
\value{
Returns a data frame (or list of data frames, depending on input) of paired TCRs
where each TCR is listed only once.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
By default, paired TCRs are listed in one row in the paired chain data
for each pairing algorithm they were called by (either T-SHELL or MAD-HYPE). This means
some TCRs are present in two rows, once for each algorithm. This
function removes duplicate TCRs and returns a data frame where each TCR pair is
found in only one row.
}
\seealso{
Other data_processing: 
\code{\link{TIRTL_process}()},
\code{\link{add_single_chain_data}()},
\code{\link{clean_pairs}()},
\code{\link{filter_duplicate_tcrs}()},
\code{\link{filter_mait}()},
\code{\link{filter_nonfunctional_TCRs}()},
\code{\link{filter_short_cdr3s}()},
\code{\link{filter_v_alleles}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()}
}
\concept{data_processing}
