% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Removes duplicate paired TCRs (by default, paired TCRs are listed twice if they
were called by both pairing algorithms)}
\usage{
remove_duplicates(data)
}
\description{
There will be duplicates in paired chain data when a pair is called by both
the T-SHELL and MAD-HYPE algorithms.
The input 'data' needs to be either a single data frame (paired chain) or a list
of data frames (paired chain)
}
\seealso{
Other data_processing: 
\code{\link{TIRTL_process}()},
\code{\link{add_single_chain_data}()},
\code{\link{clean_pairs}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()}
}
\concept{data_processing}
