% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_samples.R
\name{reorder_samples}
\alias{reorder_samples}
\title{Change the order of samples in a TIRTLseqData object}
\usage{
reorder_samples(data, samples)
}
\description{
Change the order of samples in a TIRTLseqData object
}
\seealso{
Other data_wrangling: 
\code{\link{add_metadata}()},
\code{\link{filter_dataset}()},
\code{\link{load_tirtlseq}()}
}
\concept{data_wrangling}
