% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_samples.R
\name{reorder_samples}
\alias{reorder_samples}
\title{Re-order samples in a TIRTLseqData object}
\usage{
reorder_samples(data, samples)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{samples}{the samples of 'data', in their desired order.

Either:
\enumerate{
\item A numeric vector of indices
\item A character vector of sample names
\item A character vector of metadata conditions where each entry is of the form \code{"column==value"}
}}
}
\value{
A TIRTLseqData object with re-ordered samples.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function changes the order of the samples in a TIRTLseqData object to the order that the user specifies.
}
\examples{
folder = system.file("extdata/SJTRC_TIRTL_seq_longitudinal",
  package = "TIRTLtools")
sjtrc = load_tirtlseq(folder,
  meta_columns = c("marker", "timepoint", "version"), sep = "_",
  chain = "paired", verbose = FALSE)
print(sjtrc$meta)
new_order = names(sjtrc$data) \%>\% rev()
print(new_order)
sjtrc = reorder_samples(sjtrc, new_order)
print(sjtrc$meta)

}
\seealso{
Other data_wrangling: 
\code{\link{add_metadata}()},
\code{\link{filter_dataset}()}
}
\concept{data_wrangling}
