% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_samples.R
\name{reorder_samples}
\alias{reorder_samples}
\title{Change the order of samples in a TIRTLseqData object}
\usage{
reorder_samples(data, samples)
}
\arguments{
\item{data}{a TIRTLseqData object}

\item{samples}{the samples of 'data', in their desired order. Either 1) a numeric vector of indices,
2) a character vector of sample names, or 3) a character vector of metadata
conditions where each entry is of the form \code{"column==value"}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other data_wrangling: 
\code{\link{add_metadata}()},
\code{\link{filter_dataset}()},
\code{\link{load_tirtlseq}()}
}
\concept{data_wrangling}
