% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Creates a summary table with number of reads and unique alpha/beta chains observed for each sample}
\usage{
summarize_data(data)
}
\description{
Creates a summary table with number of reads and unique alpha/beta chains observed for each sample
}
\seealso{
Other repertoire_analysis: 
\code{\link{TCRdist}()},
\code{\link{TCRdist_cpp}()},
\code{\link{calculate_diversity}()},
\code{\link{cluster_tcrs}()},
\code{\link{get_all_div_metrics}()},
\code{\link{get_all_tcrs}()},
\code{\link{get_pair_stats}()},
\code{\link{get_paired_by_read_fraction_range}()}
}
\concept{repertoire_analysis}
