% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRdist_cpp.R
\name{TCRdist_cpp}
\alias{TCRdist_cpp}
\title{Parallelized C++ implementation of TCRdist (no GPU required)}
\usage{
TCRdist_cpp(tcr1, tcr2 = NULL)
}
\arguments{
\item{tcr1}{a data frame with one TCR per row. It must have the columns "va", "vb", "cdr3a", and "cdr3b"}

\item{tcr2}{(optional) another data frame of TCRs. If supplied, TCRdist will be calculated
for every combination of one TCR from tcr1 and one TCR from tcr2. Otherwise, it will calculate TCRdist
for every pair of TCRs in tcr1.}
}
\value{
a list with two objects (or three if tcr2 is not null):
\itemize{
\item matrix - a matrix of TCRdist values
\item tcr1 - the input matrix tcr1, after pre-processing and removing unacceptable TCRs
\item tcr2 (if supplied) - the input matrix tcr2, after pre-processing and removing unacceptable TCRs
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is an alternative to the GPU version of TCRdist that is still very fast for
large datasets (tens of thousands of TCRs). It is written in C++ and will
run in parallel across available CPU cores.
}
\details{
This version of TCRdist is currently less feature-rich than \code{\link{TCRdist}()} and returns
a dense matrix as output. It does not yet allow for sparse output or writing output directly to a file.
}
\examples{
load_example_data(dataset = "SJTRC_minimal")
df = get_all_tcrs(SJTRC_minimal, chain="paired", remove_duplicates = TRUE)

result = TCRdist_cpp(df)

mat = result$matrix
node_df = result$tcr1

mat[1:5,1:5]
## note: TCRdist is symmetric. Returned matrix contains only lower-triangular values

node_df \%>\%
  mutate(alpha_nuc = paste(substr(alpha_nuc, 0, 20), "...", sep = ""),
         beta_nuc = paste(substr(beta_nuc, 0, 20), "...", sep = "")) \%>\%
  data.table::as.data.table()

}
\seealso{
Other tcr_similarity:
\code{\link[=TCRdist]{TCRdist()}},
\code{\link[=TCRdist_old]{TCRdist_old()}},
\code{\link[=TCRdist_to_igraph]{TCRdist_to_igraph()}},
\code{\link[=TCRdist_to_sparse_matrix]{TCRdist_to_sparse_matrix()}},
\code{\link[=add_to_tcr_network]{add_to_tcr_network()}},
\code{\link[=cluster_tcrs]{cluster_tcrs()}},
\code{\link[=plot_clusters]{plot_clusters()}}
}
\concept{tcr_similarity}
\keyword{internal}
