% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRdist_cpp.R
\name{TCRdist_cpp}
\alias{TCRdist_cpp}
\title{Parallelized C++ implementation of TCRdist (no GPU required)}
\usage{
TCRdist_cpp(tcr1, tcr2 = NULL, tcrdist_cutoff = 90)
}
\description{
Parallelized C++ implementation of TCRdist (no GPU required)
}
\seealso{
Other repertoire_analysis: 
\code{\link{TCRdist}()},
\code{\link{calculate_diversity}()},
\code{\link{cluster_tcrs}()},
\code{\link{get_all_div_metrics}()},
\code{\link{get_all_tcrs}()},
\code{\link{get_pair_stats}()},
\code{\link{get_paired_by_read_fraction_range}()},
\code{\link{summarize_data}()}
}
\concept{repertoire_analysis}
