% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRdist_cpp.R
\name{TCRdist_cpp}
\alias{TCRdist_cpp}
\title{Parallelized C++ implementation of TCRdist (no GPU required)}
\usage{
TCRdist_cpp(tcr1, tcr2 = NULL)
}
\arguments{
\item{tcr1}{a data frame with one TCR per row. It must have the columns "va", "vb", "cdr3a", and "cdr3b"}

\item{tcr2}{(optional) another data frame of TCRs. If supplied, TCRdist will be calculated
for every combination of one TCR from tcr1 and one TCR from tcr2. Otherwise, it will calculate TCRdist
for every pair of TCRs in tcr1.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other repertoire_analysis: 
\code{\link{TCRdist}()},
\code{\link{cluster_tcrs}()},
\code{\link{get_all_div_metrics}()},
\code{\link{get_all_tcrs}()},
\code{\link{get_pair_stats}()},
\code{\link{get_paired_by_read_fraction_range}()},
\code{\link{summarize_data}()}
}
\concept{repertoire_analysis}
