% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIRTL_process.R
\name{TIRTL_process}
\alias{TIRTL_process}
\title{Run the other data processing functions in a sequence (add_single_chain_data, clean_pairs, and identify_paired)}
\usage{
TIRTL_process(data)
}
\description{
Run the other data processing functions in a sequence (add_single_chain_data, clean_pairs, and identify_paired)
}
\seealso{
Other data_processing: 
\code{\link{add_single_chain_data}()},
\code{\link{clean_pairs}()},
\code{\link{identify_non_functional_seqs}()},
\code{\link{identify_paired}()},
\code{\link{prep_for_tcrdist}()},
\code{\link{remove_duplicates}()}
}
\concept{data_processing}
