% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIRTLtools-package.R
\docType{package}
\name{TIRTLtools-package}
\alias{TIRTLtools}
\alias{TIRTLtools-package}
\title{TIRTLtools: Analysis Tools for TIRTLseq Data}
\description{
A suite of tools for analyzing paired T-cell receptor (TCR) datasets created using the TIRTL-seq assay. Contains an efficient GPU implementation of TCRdist (Dash et al., Nature 2017).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nicholasclark.github.io/TIRTLtools/}
}

}
\author{
\strong{Maintainer}: Nicholas Clark \email{nicholas.clark00@gmail.com}

Authors:
\itemize{
  \item Mikhail Pogorelyy \email{mpogorel@fredhutch.org}
}

}
\keyword{internal}
