% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stitchr_tsv.R
\name{write_stitchr_tsv}
\alias{write_stitchr_tsv}
\title{Write a tab-separated "thimble" text file for use with Stitchr}
\usage{
write_stitchr_tsv(df, output_name, output_folder = "")
}
\description{
Write a tab-separated "thimble" text file for use with Stitchr
}
\seealso{
Other stitchr: 
\code{\link{create_thimble_df}()}
}
\concept{stitchr}
