% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stitchr_tsv.R
\name{write_stitchr_tsv}
\alias{write_stitchr_tsv}
\title{Write a tab-separated "thimble" text file for use with Stitchr}
\usage{
write_stitchr_tsv(df, output_name, output_folder = "")
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other stitchr: 
\code{\link{create_thimble_df}()}
}
\concept{stitchr}
